test(e2e): 拆分 github issue 构建回归用例
