+++ e2e/ci/github-issues.build.test.ts
@@ -1,3222 +1,11 @@
-/* eslint-disable e18e/ban-dependencies -- e2e build assertions reuse shared fs helpers to inspect generated artifacts. */
-import { originalPositionFor, TraceMap } from '@jridgewell/trace-mapping'
-import { REQUEST_GLOBAL_PRELUDE_MARKER } from '@weapp-core/constants'
-import { fs } from '@weapp-core/shared/node'
-import { execa } from 'execa'
-import { fdir } from 'fdir'
-import path from 'pathe'
-import { describe, expect, it } from 'vitest'
-import { runWeappViteBuildWithLogCapture, sanitizeBuildCommandEnv } from '../utils/buildLog'
-import { findWevuRuntimeChunk, findWevuVendorChunkContaining } from '../utils/wevu-vendor'
-
-const CLI_PATH = path.resolve(import.meta.dirname, '../../packages/weapp-vite/bin/weapp-vite.js')
-const APP_ROOT = path.resolve(import.meta.dirname, '../../e2e-apps/github-issues')
-const DIST_ROOT = path.join(APP_ROOT, 'dist')
-const ISSUE_393_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-393')
-const ISSUE_510_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-510')
-const ISSUE_595_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-595')
-const ISSUE_642_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-642')
-const ISSUE_724_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-724')
-const ISSUE_793_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-793')
-const ISSUE_838_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-838')
-const ISSUE_845_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-845')
-const ISSUE_850_DIST_ROOT = path.join(APP_ROOT, 'dist-issue-850')
-const issue826DistRoot = (jsFormat: 'cjs' | 'esm') => path.join(APP_ROOT, `dist-issue-826-${jsFormat}`)
-const SLOT_FALLBACK_COMPILER_OFF_DIST_ROOT = path.join(APP_ROOT, 'dist-slot-fallback-compiler-off')
-const SLOT_OWNER_ATTR = `__wvSlotOwnerId="{{__wvSlotOwnerId || __wvOwnerId || ''}}"`
-let standardBuildPromise: Promise<void> | null = null
-let distVariant: 'standard' | 'sourcemap' | null = null
-
-async function runBuild() {
-  if (distVariant !== 'standard') {
-    standardBuildPromise = null
-  }
-
-  standardBuildPromise ??= (async () => {
-    await fs.remove(DIST_ROOT)
-
-    await runWeappViteBuildWithLogCapture({
-      cliPath: CLI_PATH,
-      projectRoot: APP_ROOT,
-      platform: 'weapp',
-      cwd: APP_ROOT,
-      label: 'ci:github-issues',
-      skipNpm: true,
-    })
-  })()
-
-  await standardBuildPromise
-  distVariant = 'standard'
-}
-
-async function runAggregateBuild() {
-  standardBuildPromise = null
-  await fs.remove(DIST_ROOT)
-
-  await runWeappViteBuildWithLogCapture({
-    cliPath: CLI_PATH,
-    projectRoot: APP_ROOT,
-    platform: 'weapp',
-    cwd: APP_ROOT,
-    label: 'ci:github-issues:aggregate',
-    skipNpm: true,
-    env: {
-      WEAPP_VITE_E2E_TARGET_FILE: 'e2e/ide/github-issues.runtime.aggregate.test.ts',
-    },
-  })
-
-  distVariant = null
-}
-
-async function runIssue845Build() {
-  await fs.remove(ISSUE_845_DIST_ROOT)
-  await runWeappViteBuildWithLogCapture({
-    cliPath: CLI_PATH,
-    projectRoot: APP_ROOT,
-    platform: 'weapp',
-    cwd: APP_ROOT,
-    label: 'ci:github-issues:issue845',
-    skipNpm: true,
-    env: {
-      WEAPP_GITHUB_ISSUE_845_I18N: 'true',
-    },
-  })
-}
-
-async function runIssue838Build() {
-  await fs.remove(ISSUE_838_DIST_ROOT)
-  await runWeappViteBuildWithLogCapture({
-    cliPath: CLI_PATH,
-    projectRoot: APP_ROOT,
-    platform: 'weapp',
-    cwd: APP_ROOT,
-    label: 'ci:github-issues:issue838',
-    skipNpm: true,
-    env: {
-      WEAPP_GITHUB_ISSUE_838_APP_VUE_STYLES: 'true',
-    },
-  })
-}
-
-async function runIssue850Build() {
-  await fs.remove(ISSUE_850_DIST_ROOT)
-  await runWeappViteBuildWithLogCapture({
-    cliPath: CLI_PATH,
-    projectRoot: APP_ROOT,
-    platform: 'weapp',
-    cwd: APP_ROOT,
-    label: 'ci:github-issues:issue850',
-    skipNpm: true,
-    env: {
-      WEAPP_GITHUB_ISSUE_850_OUTPUT_REPLAY: 'true',
-    },
-  })
-}
-
-async function runBuildWithSourcemap() {
-  standardBuildPromise = null
-  await fs.remove(DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--sourcemap',
-  ], {
-    cwd: APP_ROOT,
-    extendEnv: false,
-    env: sanitizeBuildCommandEnv(),
-    stdio: 'inherit',
-  })
-
-  distVariant = 'sourcemap'
-}
-
-async function runBuildWithNodeEnv(nodeEnv: string) {
-  await fs.remove(DIST_ROOT)
-
-  await runWeappViteBuildWithLogCapture({
-    cliPath: CLI_PATH,
-    projectRoot: APP_ROOT,
-    platform: 'weapp',
-    cwd: APP_ROOT,
-    env: { NODE_ENV: nodeEnv },
-    label: `ci:github-issues-node-env-${nodeEnv}`,
-    skipNpm: true,
-  })
-}
-
-async function scanFiles(root: string) {
-  // eslint-disable-next-line new-cap
-  const fd = new fdir({
-    relativePaths: true,
-    pathSeparator: '/',
-  })
-  return (await fd.crawl(root).withPromise()).sort()
-}
-
-function escapeRegex(value: string) {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&')
-}
-
-function stripGeneratedScopeAttrs(code: string) {
-  return code.replace(/\sdata-v-[\w-]+=""/g, '')
-}
-
-function expectModuleReference(code: string, specifier: string) {
-  const escapedSpecifier = escapeRegex(specifier)
-  const escapedBuiltSpecifier = `(?:\\.\\.\\/)+miniprogram_npm\\/${escapedSpecifier}(?:\\/index)?`
-  expect(code).toMatch(new RegExp(`(?:require\\((['"\`])(?:${escapedSpecifier}|${escapedBuiltSpecifier})\\1\\)|from\\s+(['"\`])(?:${escapedSpecifier}|${escapedBuiltSpecifier})\\2)`))
-}
-
-function createCachedEnvLinePattern(variableName: string) {
-  return new RegExp(
-    `const ${variableName} = .*globalThis\\["__weappViteImportMetaEnv"\\].*JSON\\.parse\\(`,
-  )
-}
-
-function findGeneratedPosition(code: string, needle: string) {
-  const lines = code.split('\n')
-  const lineIndex = lines.findIndex(line => line.includes(needle))
-  expect(lineIndex).toBeGreaterThanOrEqual(0)
-  return {
-    column: lines[lineIndex].indexOf(needle),
-    line: lineIndex + 1,
-  }
-}
-
-function createObjectPropertyPattern(property: string, value: string) {
-  return new RegExp(`${escapeRegex(JSON.stringify(property))}\\s*:\\s*${escapeRegex(JSON.stringify(value))}`)
-}
-
-function expectProductionEnvironment(pageJs: string) {
-  expect(pageJs).toMatch(/issue792\s*:\s*\{\s*dev\s*:\s*false\s*,\s*mode\s*:\s*["']production["']\s*,\s*nodeEnv\s*:\s*["']production["']\s*,\s*prod\s*:\s*true\s*\}/)
-}
-
-function expectSetDataPickKeys(code: string, keys: string[]) {
-  const match = code.match(/setData\s*:\s*\{\s*pick\s*:\s*\[([\s\S]*?)\]/)
-  expect(match?.[1]).toBeTruthy()
-  for (const key of keys) {
-    expect(match![1]).toContain(JSON.stringify(key))
-  }
-}
-
-function expectSetDataPickOmitsKeys(code: string, keys: string[]) {
-  const match = code.match(/setData\s*:\s*\{\s*pick\s*:\s*\[([\s\S]*?)\]/)
-  expect(match?.[1]).toBeTruthy()
-  for (const key of keys) {
-    expect(match![1]).not.toContain(JSON.stringify(key))
-  }
-}
-
-function createSerializedJsonEntryPattern(key: string, value: string) {
-  return new RegExp(escapeRegex(`\\"${key}\\":\\"${value}\\"`))
-}
-
-function resolveSharedRuntimeImport(sourceFilePath: string, sourceCode: string) {
-  const relativeImports = [
-    ...[...sourceCode.matchAll(/require\((['"`])([^"'`]+)\1\)/g)].map(match => match[2]!),
-    ...[...sourceCode.matchAll(/from\s+(['"`])([^"'`]+)\1/g)].map(match => match[2]!),
-  ]
-  const candidates = relativeImports
-    .map(importee => path.resolve(path.dirname(sourceFilePath), importee))
-    .filter(candidate => candidate.startsWith(DIST_ROOT) && path.extname(candidate) === '.js')
-
-  return candidates.find(candidate => candidate.includes(`${path.sep}weapp-vendors${path.sep}`))
-    ?? candidates[0]
-}
-
-async function runIssue393Build() {
-  await fs.remove(ISSUE_393_DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...process.env,
-      WEAPP_GITHUB_ISSUE_393: 'true',
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
-
-async function runIssue510AugmentedBuild() {
-  await fs.remove(ISSUE_510_DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_ISSUE_510_AUGMENTED: 'true',
-    },
-  })
-
-  distVariant = null
-}
-
-async function runIssue547AugmentedBuild() {
-  await fs.remove(DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_ISSUE_547_AUGMENTED: 'true',
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
-
-async function runIssue558AugmentedBuild() {
-  await fs.remove(DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_ISSUE_558_AUGMENTED: 'true',
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
-
-async function runIssue595ScopedBuild() {
-  await fs.remove(ISSUE_595_DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--scope',
-    'main,subpackages/item',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_ISSUE_595_SCOPED: 'true',
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
-
-async function runIssue615AugmentedBuild() {
-  await fs.remove(DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_ISSUE_615_AUGMENTED: 'true',
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
-
-async function runIssue621AugmentedBuild() {
-  await fs.remove(DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_ISSUE_621_AUGMENTED: 'true',
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
-
-async function runIssue642Build() {
-  standardBuildPromise = null
-  await fs.remove(ISSUE_642_DIST_ROOT)
-
-  await runWeappViteBuildWithLogCapture({
-    cliPath: CLI_PATH,
-    projectRoot: APP_ROOT,
-    platform: 'weapp',
-    cwd: APP_ROOT,
-    label: 'ci:github-issues:issue642',
-    outDir: 'dist-issue-642',
-    skipNpm: true,
-    env: {
-      WEAPP_VITE_E2E_TARGET_FILE: 'e2e/ide/github-issues.runtime.issue642.test.ts',
-      WEAPP_GITHUB_ISSUE_642_SCOPED: 'true',
-    },
-  })
-
-  distVariant = null
-}
-
-async function runIssue724Build() {
-  standardBuildPromise = null
-  await fs.remove(ISSUE_724_DIST_ROOT)
-
-  await runWeappViteBuildWithLogCapture({
-    cliPath: CLI_PATH,
-    projectRoot: APP_ROOT,
-    platform: 'weapp',
-    cwd: APP_ROOT,
-    label: 'ci:github-issues:issue724',
-    outDir: 'dist-issue-724',
-    skipNpm: true,
-    env: {
-      WEAPP_GITHUB_ISSUE_724_PROBE: 'true',
-    },
-  })
-
-  distVariant = null
-}
-
-interface AppJsonWithSubPackages {
-  pages?: string[]
-  subPackages?: Array<{ root?: string, pages?: string[] }>
-  subpackages?: Array<{ root?: string, pages?: string[] }>
-}
-
-async function runIssue564AugmentedBuild() {
-  await fs.remove(DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_ISSUE_564_AUGMENTED: 'true',
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
-
-async function runSlotFallbackCompilerOffBuild() {
-  await fs.remove(SLOT_FALLBACK_COMPILER_OFF_DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_SLOT_FALLBACK_COMPILER_OFF: 'true',
-    },
-  })
-
-  distVariant = null
-}
-
-async function runIssue793Build(scope: 'main' | 'main-with-subpackage' | 'subpackage' = 'main') {
-  await fs.remove(ISSUE_793_DIST_ROOT)
-
-  await execa('node', [
-    CLI_PATH,
-    'build',
-    APP_ROOT,
-    '--platform',
-    'weapp',
-    '--skipNpm',
-    '--config',
-    path.join(APP_ROOT, 'weapp-vite.config.ts'),
-  ], {
-    stdio: 'inherit',
-    env: {
-      ...sanitizeBuildCommandEnv(),
-      WEAPP_GITHUB_ISSUE_793_BUILD_SCOPE: scope === 'main' ? 'true' : scope,
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
-
-async function runIssue826Build(jsFormat: 'cjs' | 'esm') {
-  const distRoot = issue826DistRoot(jsFormat)
-  await fs.remove(distRoot)
-
-  await runWeappViteBuildWithLogCapture({
-    cliPath: CLI_PATH,
-    projectRoot: APP_ROOT,
-    platform: 'weapp',
-    cwd: APP_ROOT,
-    jsFormat,
-    label: `ci:github-issues:issue826:${jsFormat}`,
-    outDir: `dist-issue-826-${jsFormat}`,
-    skipNpm: true,
-    env: {
-      WEAPP_GITHUB_ISSUE_826_PRESERVE: 'true',
-      WEAPP_GITHUB_ISSUE_826_OUT_DIR: `dist-issue-826-${jsFormat}`,
-    },
-  })
-
-  standardBuildPromise = null
-  distVariant = null
-}
+import { describe } from 'vitest'
+import { registerGithubIssuesBuildCases } from './githubIssuesBuild/cases'
+import {
+  createGithubIssuesBuildCaseContext,
+  registerGithubIssuesBuildLegacyCases,
+} from './githubIssuesBuild/legacy'
 
 describe.sequential('e2e app: github-issues (build)', () => {
-  it('issue #852: normalizes JavaScript numeric literals in WXML expressions', async () => {
-    await runBuild()
-
-    const wxml = await fs.readFile(path.join(DIST_ROOT, 'pages/issue-852/index.wxml'), 'utf8')
-
-    expect(wxml).toContain('data-decimal="{{1000000000000}}"')
-    expect(wxml).toContain('data-fraction="{{1050.95}}"')
-    expect(wxml).toContain('data-binary="{{41349}}"')
-    expect(wxml).toContain('data-octal="{{1198}}"')
-    expect(wxml).toContain('data-hex="{{10531008}}"')
-    expect(wxml).toContain(`data-bigint="{{'1000000000000000000000'}}"`)
-    expect(wxml).toContain('{{81985529216486900}}')
-    expect(wxml).not.toMatch(/(?:\d_\d|0[bxo]|\d+n\b)/i)
-  })
-
-  it('PR #838: injects explicitly included shared styles into app.vue', async () => {
-    await runIssue838Build()
-
-    const appStyle = await fs.readFile(path.join(ISSUE_838_DIST_ROOT, 'app.wxss'), 'utf8')
-    const sharedStyle = await fs.readFile(
-      path.join(ISSUE_838_DIST_ROOT, 'styles/issue-838-app-vue.wxss'),
-      'utf8',
-    )
-
-    expect(appStyle).toContain('@import \'./styles/issue-838-app-vue.wxss\';')
-    expect(sharedStyle).toContain('color: #123456;')
-  })
-
-  it('issue #850: processes completed independent outputs only once', async () => {
-    await runIssue850Build()
-
-    const independentStyle = await fs.readFile(
-      path.join(ISSUE_850_DIST_ROOT, 'subpackages/issue-850/index.wxss'),
-      'utf8',
-    )
-    const appJson = await fs.readJSON(path.join(ISSUE_850_DIST_ROOT, 'app.json')) as {
-      subPackages?: Array<{ independent?: boolean, root?: string }>
-    }
-
-    expect(independentStyle.match(/issue-850-output-marker/g)).toHaveLength(1)
-    expect(appJson.subPackages).toContainEqual(expect.objectContaining({
-      independent: true,
-      root: 'subpackages/issue-850',
-    }))
-  })
-
-  it('issue #845: emits first-party i18n assets for native and Vue templates', async () => {
-    await runIssue845Build()
-
-    const nativeWxml = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'pages/issue-845-native/index.wxml'), 'utf8')
-    const nativeComponentWxml = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'components/issue-845-i18n-card/index.wxml'), 'utf8')
-    const vueJs = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'pages/issue-845-vue/index.js'), 'utf8')
-    const vueWxml = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'pages/issue-845-vue/index.wxml'), 'utf8')
-    const localesWxs = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'i18n/locales.wxs'), 'utf8')
-    const localesJs = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'i18n/locales.js'), 'utf8')
-    const runtimePath = resolveSharedRuntimeImport(path.join(ISSUE_845_DIST_ROOT, 'i18n/locales.js'), localesJs)
-    expect(runtimePath).toBeTruthy()
-    const runtimeJs = await fs.readFile(runtimePath!, 'utf8')
-    const ordinaryWxml = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'subpackages/issue-845-normal/index.wxml'), 'utf8')
-    const independentWxml = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'subpackages/issue-845-independent/index.wxml'), 'utf8')
-    const independentJs = await fs.readFile(path.join(ISSUE_845_DIST_ROOT, 'subpackages/issue-845-independent/index.js'), 'utf8')
-    const independentRoot = path.join(ISSUE_845_DIST_ROOT, 'subpackages/issue-845-independent')
-    const independentRuntimeEntryPath = path.join(independentRoot, 'i18n/locales.js')
-    const independentRuntimeEntryJs = await fs.readFile(independentRuntimeEntryPath, 'utf8')
-    const independentRuntimePath = resolveSharedRuntimeImport(independentRuntimeEntryPath, independentRuntimeEntryJs)
-    expect(independentRuntimePath).toBeTruthy()
-    const independentRuntimeJs = await fs.readFile(independentRuntimePath!, 'utf8')
-    const appJson = await fs.readJSON(path.join(ISSUE_845_DIST_ROOT, 'app.json')) as {
-      subPackages?: Array<{ independent?: boolean, root?: string }>
-    }
-
-    expect(nativeWxml).toContain('<wxs module="i18n" src="../../i18n/locales.wxs"/>')
-    expect(nativeWxml).toContain('i18n.t(__wv_i18n_locale, \'issue845.greeting\', { user })')
-    expect(nativeComponentWxml).toContain('i18n.t(__wv_i18n_locale, \'issue845.greeting\', { user })')
-    expect(vueJs).toContain('__WEAPP_VITE_I18N__')
-    expect(vueJs).not.toContain('behaviors: [{}]')
-    expect(vueWxml).toContain('i18n.t(__wv_i18n_locale, \'issue845.greeting\', { user })')
-    expect(ordinaryWxml).toContain('<wxs module="i18n" src="./i18n/locales.wxs"/>')
-    expect(independentWxml).toContain('<wxs module="i18n" src="./i18n/locales.wxs"/>')
-    expect(appJson.subPackages).toContainEqual(expect.objectContaining({
-      independent: true,
-      root: 'subpackages/issue-845-independent',
-    }))
-    expect(independentJs).not.toContain('require("../../')
-    expect(path.relative(independentRoot, independentRuntimePath!)).not.toMatch(/^\.\.(?:[/\\]|$)/)
-    expect(independentRuntimeJs).toContain('RangeError')
-    expect(independentRuntimeJs).toContain('__WEAPP_VITE_I18N__')
-    expect(localesWxs).toContain('Fallback text')
-    expect(runtimeJs).toContain('RangeError')
-    for (const root of ['subpackages/issue-845-normal', 'subpackages/issue-845-independent']) {
-      expect(await fs.pathExists(path.join(ISSUE_845_DIST_ROOT, root, 'i18n/locales.js'))).toBe(true)
-      expect(await fs.pathExists(path.join(ISSUE_845_DIST_ROOT, root, 'i18n/locales.wxs'))).toBe(true)
-    }
-  })
-
-  it('issue #826: preserves configured source module paths for debugging', async () => {
-    for (const jsFormat of ['cjs', 'esm'] as const) {
-      await runIssue826Build(jsFormat)
-      const distRoot = issue826DistRoot(jsFormat)
-
-      const files = await scanFiles(distRoot)
-      expect(files).toEqual(expect.arrayContaining([
-        'issue-fixtures/issue-826/services/shared.js',
-        'issue-fixtures/issue-826/services/single.js',
-        'issue-fixtures/issue-826/utils/barrel.js',
-        'issue-fixtures/issue-826/utils/leafA.js',
-        'issue-fixtures/issue-826/utils/leafB.js',
-        'issue-fixtures/issue-826/utils/shared.js',
-        'issue-fixtures/issue-826/utils/single.js',
-      ]))
-      expect(files).not.toContain('pages/issue-826/local.js')
-
-      const pageCode = await fs.readFile(path.join(distRoot, 'pages/issue-826/index.js'), 'utf8')
-      expect(pageCode).toContain('__ISSUE_826_LOCAL_UNMATCHED__')
-      expect(pageCode).not.toContain('__ISSUE_826_UTIL_SINGLE__')
-      expect(pageCode).not.toContain('__ISSUE_826_SERVICE_SINGLE__')
-      expect(pageCode).toMatch(/issue-fixtures\/issue-826\/utils\/single\.js/)
-      expect(pageCode).toMatch(/issue-fixtures\/issue-826\/services\/single\.js/)
-
-      const barrelCode = await fs.readFile(path.join(distRoot, 'issue-fixtures/issue-826/utils/barrel.js'), 'utf8')
-      expect(barrelCode).not.toContain('__ISSUE_826_UTIL_LEAF_A__')
-      expect(barrelCode).not.toContain('__ISSUE_826_UTIL_LEAF_B__')
-      expect(barrelCode).toMatch(/(?:require\((['"`])\.\/leafA\.js\1\)|from\s+(['"`])\.\/leafA\.js\2)/)
-      expect(barrelCode).toMatch(/(?:require\((['"`])\.\/leafB\.js\1\)|from\s+(['"`])\.\/leafB\.js\2)/)
-
-      const secondPageCode = await fs.readFile(path.join(distRoot, 'pages/issue-826/second.js'), 'utf8')
-      expect(secondPageCode).not.toContain('__ISSUE_826_UTIL_SHARED__')
-      expect(secondPageCode).toMatch(/issue-fixtures\/issue-826\/utils\/shared\.js/)
-      await fs.remove(distRoot)
-    }
-  })
-
-  it('issue #793: keeps derived app config aligned with the scoped routes', async () => {
-    await runIssue793Build()
-
-    const appJson = await fs.readJSON(path.join(ISSUE_793_DIST_ROOT, 'app.json')) as {
-      entryPagePath?: string
-      pages?: string[]
-      preloadRule?: Record<string, unknown>
-      subPackages?: Array<{ root?: string, pages?: string[] }>
-      tabBar?: { list?: Array<{ pagePath?: string }> }
-    }
-
-    expect(appJson.entryPagePath).toBe('pages/issue-793/index')
-    expect(appJson.pages).toHaveLength(2)
-    expect(appJson.pages).toEqual(expect.arrayContaining([
-      'pages/issue-793/index',
-      'pages/issue-793-settings/index',
-    ]))
-    expect(appJson.preloadRule).toBeUndefined()
-    expect(appJson.subPackages).toEqual([])
-    expect(appJson.tabBar?.list?.map(item => item.pagePath)).toEqual([
-      'pages/issue-793/index',
-      'pages/issue-793-settings/index',
-    ])
-    expect(await fs.pathExists(path.join(ISSUE_793_DIST_ROOT, 'subs'))).toBe(false)
-
-    await runIssue793Build('subpackage')
-
-    const subpackageAppJson = await fs.readJSON(path.join(ISSUE_793_DIST_ROOT, 'app.json')) as {
-      entryPagePath?: string
-      pages?: string[]
-      preloadRule?: Record<string, unknown>
-      subPackages?: Array<{ root?: string, pages?: string[] }>
-      tabBar?: unknown
-    }
-
-    expect(subpackageAppJson.entryPagePath).toBeUndefined()
-    expect(subpackageAppJson.pages).toEqual([])
-    expect(subpackageAppJson.preloadRule).toBeUndefined()
-    expect(subpackageAppJson.subPackages).toEqual([
-      { root: 'subs', pages: ['issue-793/index'] },
-    ])
-    expect(subpackageAppJson.tabBar).toBeUndefined()
-    expect(await fs.pathExists(path.join(ISSUE_793_DIST_ROOT, 'pages'))).toBe(false)
-
-    await runIssue793Build('main-with-subpackage')
-
-    const mainWithSubpackageAppJson = await fs.readJSON(path.join(ISSUE_793_DIST_ROOT, 'app.json')) as {
-      pages?: string[]
-      preloadRule?: Record<string, { packages?: string[] }>
-      subPackages?: Array<{ root?: string, pages?: string[] }>
-      tabBar?: { list?: Array<{ pagePath?: string }> }
-    }
-
-    expect(mainWithSubpackageAppJson.pages).toEqual(expect.arrayContaining([
-      'pages/issue-793/index',
-      'pages/issue-793-settings/index',
-    ]))
-    expect(mainWithSubpackageAppJson.subPackages).toEqual([
-      { root: 'subs', pages: ['issue-793/index'] },
-    ])
-    expect(mainWithSubpackageAppJson.preloadRule).toEqual({
-      'pages/issue-793/index': {
-        packages: ['subs'],
-      },
-      'subs/issue-793/index': {
-        packages: ['__APP__'],
-      },
-    })
-    expect(mainWithSubpackageAppJson.tabBar?.list?.map(item => item.pagePath)).toEqual([
-      'pages/issue-793/index',
-      'pages/issue-793-settings/index',
-    ])
-    expect(await fs.pathExists(path.join(ISSUE_793_DIST_ROOT, 'pages/issue-793/index.js'))).toBe(true)
-    expect(await fs.pathExists(path.join(ISSUE_793_DIST_ROOT, 'subs/issue-793/index.js'))).toBe(true)
-  })
-
-  it('issue #805: lowers optional chaining inside nullish coalescing expressions', async () => {
-    await runBuild()
-
-    const pageWxml = await fs.readFile(path.join(DIST_ROOT, 'pages/issue-805/index.wxml'), 'utf-8')
-
-    expect(pageWxml).not.toContain('?.')
-    expect(pageWxml).not.toContain('??')
-    expect(pageWxml).toContain('a==null?undefined:a.b==null?undefined:a.b.d')
-    expect(pageWxml).toContain('!=null?')
-    expect(pageWxml).toContain(':4')
-  })
-
-  it('issue #724: keeps Vue SFC script, template, and style requests isolated', async () => {
-    await runIssue724Build()
-
-    const pageBase = path.join(ISSUE_724_DIST_ROOT, 'pages/issue-724/index')
-    const componentBase = path.join(ISSUE_724_DIST_ROOT, 'components/issue-724/RoutingProbe/index')
-    const pageJs = await fs.readFile(`${pageBase}.js`, 'utf-8')
-    const pageWxml = await fs.readFile(`${pageBase}.wxml`, 'utf-8')
-    const pageWxss = await fs.readFile(`${pageBase}.wxss`, 'utf-8')
-    const pageJson = await fs.readJSON(`${pageBase}.json`) as { usingComponents?: Record<string, string> }
-    const componentJs = await fs.readFile(`${componentBase}.js`, 'utf-8')
-    const componentWxml = await fs.readFile(`${componentBase}.wxml`, 'utf-8')
-    const componentWxss = await fs.readFile(`${componentBase}.wxss`, 'utf-8')
-
-    expect(pageJs).toContain('issue-724-script-marker')
-    expect(pageWxml).toContain('issue-724-template-marker')
-    expect(pageWxml).toContain('<issue724-routing-probe />')
-    expect(pageWxss).toContain('.issue-724-page')
-    expect(pageWxss).toContain('.issue-724-template-marker')
-    expect(pageWxss).toContain('.issue-724-style-src')
-    expect(pageWxss).not.toContain('.issue-724-component')
-    expect(pageJson.usingComponents).toMatchObject({
-      'issue724-routing-probe': '/components/issue-724/RoutingProbe/index',
-    })
-
-    expect(componentJs).toContain('issue-724-component-script-marker')
-    expect(componentWxml).toContain('issue-724-component-template-marker')
-    expect(componentWxss).toContain('.issue-724-component')
-    expect(componentWxss).not.toContain('.issue-724-page')
-
-    for (const output of [pageJs, pageWxml, pageWxss, componentJs, componentWxml, componentWxss]) {
-      expect(output).not.toMatch(/<(?:template|script|style)(?:\s|>)/)
-    }
-  })
-
-  it('issue #764: installs axios request globals before page modules execute', async () => {
-    await runBuild()
-
-    const pageJs = await fs.readFile(path.join(DIST_ROOT, 'pages/issue-764/index.js'), 'utf8')
-    const appPreludeJs = await fs.readFile(path.join(DIST_ROOT, 'app.prelude.js'), 'utf8')
-
-    expect(pageJs).toContain('app.prelude.js')
-    expect(appPreludeJs).toContain(`/* ${REQUEST_GLOBAL_PRELUDE_MARKER} */`)
-    expect(appPreludeJs).toContain('"fetch"')
-    expect(appPreludeJs).toContain('"XMLHttpRequest"')
-    expect(appPreludeJs).toContain('"URL"')
-  })
-
-  it('issue #806: emits PascalCase Vue components as kebab-case WXML tags and registrations', async () => {
-    await runBuild()
-
-    const pageWxml = await fs.readFile(path.join(DIST_ROOT, 'pages/issue-806/index.wxml'), 'utf8')
-    const pageJson = await fs.readJSON(path.join(DIST_ROOT, 'pages/issue-806/index.json')) as {
-      usingComponents?: Record<string, string>
-    }
-
-    expect(pageWxml).toContain('<pascal-card vue-slots="{{ {default:true} }}">issue-806</pascal-card>')
-    expect(pageWxml).not.toContain('<PascalCard')
-    expect(pageJson.usingComponents).toMatchObject({
-      'pascal-card': '/components/issue-806/PascalCard/index',
-    })
-    expect(pageJson.usingComponents).not.toHaveProperty('PascalCard')
-  })
-
-  it('discussion #338: emits mapped wxml tags from vue html-style templates', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-338/index.wxml')
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-
-    expect(pageWxml).toContain(`<weapp-layout-default ${SLOT_OWNER_ATTR}>`)
-    expect(pageWxml).toContain('<block wx:if="{{true}}"><view')
-    expect(pageWxml).toContain('class="div issue338-page"')
-    expect(pageWxml).toContain('id="issue338-page"')
-    expect(pageWxml).toContain('data-e2e-issue="338"')
-    expect(pageWxml).toContain('<text class="span issue338-title">{{title}}</text>')
-    expect(pageWxml).toContain('<image class="img issue338-cover" src="{{cover}}" mode="aspectFit" />')
-    expect(pageWxml).toContain('<view class="section issue338-links">')
-    expect(pageWxml).toContain('<navigator class="a issue338-link" wx:for="{{links}}"')
-    expect(pageWxml).toContain('url="{{\'/pages/\'+link+\'/index\'}}"')
-    expect(pageWxml).not.toContain('<div')
-    expect(pageWxml).not.toContain('<span')
-    expect(pageWxml).not.toContain('<img')
-    expect(pageWxml).not.toContain('<section')
-    expect(pageWxml).not.toContain('<a ')
-  })
-
-  it('issue #563: applies app.vue template as app shell without app.wxml', async () => {
-    await runBuild()
-
-    const pageWxml = await fs.readFile(path.join(DIST_ROOT, 'pages/issue-338/index.wxml'), 'utf-8')
-    const pageJson = await fs.readJson(path.join(DIST_ROOT, 'pages/issue-338/index.json')) as {
-      usingComponents?: Record<string, string>
-    }
-    const appShellWxml = await fs.readFile(path.join(DIST_ROOT, '__weapp_vite_app_shell.wxml'), 'utf-8')
-    const appShellJson = await fs.readJson(path.join(DIST_ROOT, '__weapp_vite_app_shell.json')) as Record<string, unknown>
-
-    expect(await fs.pathExists(path.join(DIST_ROOT, 'app.wxml'))).toBe(false)
-    expect(appShellWxml).toContain('issue-563-app-shell')
-    expect(appShellWxml).toContain('<slot />')
-    expect(appShellWxml).not.toContain('scoped-slots-default')
-    expect(appShellJson).toEqual({
-      component: true,
-      styleIsolation: 'apply-shared',
-    })
-    expect(pageWxml).toContain(`<weapp-app-shell ${SLOT_OWNER_ATTR}><weapp-layout-default ${SLOT_OWNER_ATTR}>`)
-    expect(pageWxml).toContain('</weapp-layout-default></weapp-app-shell>')
-    expect(pageJson.usingComponents).toMatchObject({
-      'weapp-app-shell': '/__weapp_vite_app_shell',
-      'weapp-layout-default': '/layouts/default',
-    })
-  })
-
-  it('issue #431: replaces import.meta.env expressions inside native wxml files', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-431/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-431/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('{{\'issue-431 native wxml env replacement\'}}')
-    expect(pageWxml).toContain('{{\'https://static.example.com/issue-431\'}}/logo.png')
-    expect(pageWxml).toContain('data-case="double"')
-    expect(pageWxml).toContain('data-template-url="{{\'/pages/issue-431/index.wxml\'}}"')
-    expect(pageWxml).toContain('data-template-dir="{{\'/pages/issue-431\'}}"')
-    expect(pageWxml).toContain('data-env-label="{{\'issue-431 native wxml env replacement\'}}"')
-    expect(pageWxml).toContain('data-env-base="{{\'https://static.example.com/issue-431\'}}"')
-    expect(pageWxml).toContain('data-case=\'single\'')
-    expect(pageWxml).toContain('data-template-url=\'{{"/pages/issue-431/index.wxml"}}\'')
-    expect(pageWxml).toContain('data-template-dir=\'{{"/pages/issue-431"}}\'')
-    expect(pageWxml).toContain('data-env-label=\'{{"issue-431 native wxml env replacement"}}\'')
-    expect(pageWxml).toContain('data-env-base=\'{{"https://static.example.com/issue-431"}}\'')
-    expect(pageWxml).not.toContain('import.meta.env')
-    expect(pageWxml).not.toContain('import.meta.url')
-    expect(pageWxml).not.toContain('import.meta.dirname')
-    expect(pageJs).toContain('/pages/issue-431/index.js')
-    expect(pageJs).toContain('/pages/issue-431')
-    expect(pageJs).toContain('importMetaSnapshot')
-    expect(pageJs).toMatch(createObjectPropertyPattern('url', '/pages/issue-431/index.js'))
-    expectProductionEnvironment(pageJs)
-    expect(pageJs).not.toContain('import.meta.url')
-    expect(pageJs).not.toContain('import.meta.dirname')
-    expect(pageJs).not.toContain('import.meta.env')
-  })
-
-  it('issue #792: production build overrides an inherited development NODE_ENV', async () => {
-    await runBuildWithNodeEnv('development')
-
-    const pageJs = await fs.readFile(path.join(DIST_ROOT, 'pages/issue-431/index.js'), 'utf-8')
-
-    expectProductionEnvironment(pageJs)
-  })
-
-  it('issue #429: prefers local component auto import and warns when name collides with a builtin component tag', async () => {
-    await runBuild()
-
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-429/index.json')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-429/index.wxml')
-    const pageJson = await fs.readJson(pageJsonPath)
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-429 builtin-name auto import')
-    expect(pageJson.usingComponents).toMatchObject({
-      'list-view': '/components/issue-429/list-view/index',
-    })
-  })
-
-  it('issue #520: injects slot metadata for resolver-imported wevu components', async () => {
-    await runBuild()
-
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-520/index.json')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-520/index.wxml')
-    const pageJson = await fs.readJson(pageJsonPath)
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-
-    expect(pageJson.usingComponents).toMatchObject({
-      'issue520-resolver-slot-card': '/components/issue-520/ResolverSlotCard/index',
-    })
-    expect(pageWxml).toContain('issue520-resolver-slot-card vue-slots="{{ {header:true,default:true} }}"')
-    expect(pageWxml).not.toContain('vue-slots="{{{')
-    expect(pageWxml).not.toContain('vue-slots="{{ {[')
-    expect(pageWxml).toContain('issue-520 resolver slot header')
-    expect(pageWxml).toContain('issue-520 resolver slot default')
-  })
-
-  it('issue #651: resolves custom resolver SFC entries without relying on .vue suffix in resolvedId', async () => {
-    await runBuild()
-
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-651/index.json')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-651/index.wxml')
-    const noExtComponentWxmlPath = path.join(DIST_ROOT, 'issue-fixtures/issue-651/ResolverNoExt/index.wxml')
-    const noExtComponentWxssPath = path.join(DIST_ROOT, 'issue-fixtures/issue-651/ResolverNoExt/index.wxss')
-    const withExtComponentWxssPath = path.join(DIST_ROOT, 'issue-fixtures/issue-651/ResolverWithExt/index.wxss')
-    const pageJson = await fs.readJson(pageJsonPath) as { usingComponents?: Record<string, string> }
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const noExtComponentWxml = await fs.readFile(noExtComponentWxmlPath, 'utf-8')
-    const noExtComponentWxss = await fs.readFile(noExtComponentWxssPath, 'utf-8')
-    const withExtComponentWxss = await fs.readFile(withExtComponentWxssPath, 'utf-8')
-
-    expect(pageJson.usingComponents).toMatchObject({
-      'issue651-resolver-no-ext': '/issue-fixtures/issue-651/ResolverNoExt/index',
-      'issue651-resolver-with-ext': '/issue-fixtures/issue-651/ResolverWithExt/index',
-    })
-    expect(pageWxml).toContain('issue651-resolver-no-ext vue-slots="{{ {default:true} }}"')
-    expect(pageWxml).not.toContain('vue-slots="{{{')
-    expect(pageWxml).not.toContain('vue-slots="{{ {[')
-    expect(pageWxml).toContain('issue651-resolver-with-ext')
-    expect(noExtComponentWxml).toContain('<slot />')
-    expect(noExtComponentWxss).toContain('.issue651-no-ext')
-    expect(withExtComponentWxss).toContain('.issue651-with-ext')
-  })
-
-  it('issue #528: compiles slot fallback content to conditional wxml branches', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-528/index.wxml')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-528/SlotFallbackCard/index.wxml')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue528-slot-fallback-card class="issue528-card-provided" vue-slots="{{ {header:true,default:true} }}"')
-    expect(pageWxml).not.toContain('vue-slots="{{{')
-    expect(pageWxml).not.toContain('vue-slots="{{ {[')
-    expect(pageWxml).not.toContain('vue-slot-flags')
-    expect(componentWxml).toContain(`<block wx:if="{{vueSlots&&vueSlots.header}}">`)
-    expect(componentWxml).toContain('<slot name="header" /></block><block wx:else><text class="issue528-fallback-header">issue-528 fallback header</text></block>')
-    expect(componentWxml).toContain(`<block wx:if="{{vueSlots&&vueSlots.default}}">`)
-    expect(componentWxml).toContain('<slot /></block><block wx:else><text class="issue528-fallback-default">{{fallbackDefault}}</text></block>')
-    expect(componentWxml).not.toContain('vueSlots[')
-    expect(componentWxml).not.toContain('<slot name="header"><text class="issue528-fallback-header">')
-    expect(componentWxml).not.toContain('<slot><text class="issue528-fallback-default">')
-  })
-
-  it('issue #530: compiles default slot fallback guards without slot array scans', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-530/index.wxml')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-530/SlotFallbackProbe/index.wxml')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-
-    const providedProbe = pageWxml
-      .match(/<issue530-slot-fallback-probe[^>]*\/>/g)
-      ?.find(tag => tag.includes('class="issue530-card-provided"'))
-    expect(providedProbe).toBeDefined()
-    expect(providedProbe!).toContain('vue-slots=')
-    expect(providedProbe!).toContain('generic:scoped-slots-default=')
-    expect(componentWxml).toContain('<block wx:if="{{vueSlots&&vueSlots.default}}">')
-    expect(componentWxml).toContain('<scoped-slots-default wx:if="{{__wvSlotOwnerId}}"')
-    expect(componentWxml).toContain('<block wx:else><text class="issue530-fallback-default">issue-530 fallback default</text><text class="issue530-scoped-fallback-default">issue-530 scoped fallback default</text></block>')
-    expect(componentWxml).not.toContain('vueSlots[')
-  })
-
-  it('scoped slot outlet fallback: keeps native named slot projection beside scoped generic outlet', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/scoped-slot-outlet-fallback/index.wxml')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/scoped-slot-outlet-fallback/BackList/index.wxml')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('back-list class="scoped-slot-outlet-fallback-card"')
-    expect(pageWxml).toContain('slot="main"')
-    expect(pageWxml).toContain('slot="footer"')
-    expect(pageWxml).not.toContain('generic:scoped-slots-main=')
-    expect(componentWxml).toContain('<scoped-slots-main wx:if="{{__wvSlotOwnerId}}"')
-    expect(componentWxml).toContain('<block wx:if="{{!__wvSlotOwnerId}}"><slot name="main" /></block>')
-    expect(componentWxml).toContain('<scoped-slots-footer wx:if="{{__wvSlotOwnerId}}"')
-    expect(componentWxml).toContain('<block wx:if="{{!__wvSlotOwnerId}}"><slot name="footer" /></block>')
-    expect(componentWxml).toContain(`__wvSlotProps="{{['list',back.state.list]}}"`)
-    expect(componentWxml).not.toContain('<slot name="main" /><scoped-slots-main')
-  })
-
-  it('slot fallback compiler off: keeps plain fallback guards independent from scoped slot compiler', async () => {
-    await runSlotFallbackCompilerOffBuild()
-
-    const pageWxmlPath = path.join(SLOT_FALLBACK_COMPILER_OFF_DIST_ROOT, 'pages/slot-fallback-compiler-off/index.wxml')
-    const componentWxmlPath = path.join(SLOT_FALLBACK_COMPILER_OFF_DIST_ROOT, 'components/slot-fallback-compiler-off/PlainSlotFallbackCard/index.wxml')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('plain-slot-fallback-card class="slot-fallback-off-card-provided" vue-slots="{{ {header:true,default:true} }}"')
-    expect(pageWxml).not.toContain('vue-slots="{{{')
-    expect(pageWxml).not.toContain('vue-slots="{{ {[')
-    expect(pageWxml).not.toContain('generic:scoped-slots-')
-    expect(componentWxml).toContain(`<block wx:if="{{vueSlots&&vueSlots.header}}">`)
-    expect(componentWxml).toContain('<slot name="header" />')
-    expect(componentWxml).toContain('class="slot-fallback-off-fallback-header" data-slot-fallback-off-case="fallback-header"')
-    expect(componentWxml).toContain('slot-fallback-compiler-off fallback header')
-    expect(componentWxml).toContain(`<block wx:if="{{vueSlots&&vueSlots.default}}">`)
-    expect(componentWxml).toContain('<slot />')
-    expect(componentWxml).toContain('class="slot-fallback-off-fallback-default" data-slot-fallback-off-case="fallback-default"')
-    expect(componentWxml).toContain('{{fallbackDefault}}')
-    expect(componentWxml).not.toContain('<slot name="header"><text class="slot-fallback-off-fallback-header">')
-    expect(componentWxml).not.toContain('<slot><text class="slot-fallback-off-fallback-default">')
-  })
-
-  it('issue #550: exposes current route name through useRoute()', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-550/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-550/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('data-route-name="{{routeName}}"')
-    expect(pageWxml).toContain('data-matched-name="{{routeMatchedName}}"')
-    expect(pageJs).toContain('pages/issue-550/index')
-    expect(pageJs).toContain('route.name')
-  })
-
-  it('issue #705: compiles router route sync runtime probe', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-705/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-705/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('data-route-path="{{routePath}}"')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('pages/issue-550/index')
-  })
-
-  it('issue #553: maps component v-model arguments to the matching prop and update event', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-553/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-553/index.js')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-553/ModelArgumentProbe/index.wxml')
-    const componentJsPath = path.join(DIST_ROOT, 'components/issue-553/ModelArgumentProbe/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const componentJs = await fs.readFile(componentJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('<model-argument-probe')
-    expect(pageWxml).toContain('abc="{{abc}}"')
-    expect(pageWxml).toContain('modelValue="{{modelValue}}"')
-    expect(pageWxml).toContain('bind:update-abc="__weapp_vite_inline"')
-    expect(pageWxml).toContain('bind:update-modelvalue="__weapp_vite_inline"')
-    expect(pageWxml).not.toContain('abc="{{modelValue}}"')
-    expect(pageWxml).not.toContain('modelValue="{{abc}}"')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('abc-seed')
-    expect(pageJs).toContain('model-seed')
-    expect(componentWxml).toContain('child abc = {{abcModel}}')
-    expect(componentWxml).toContain('child model = {{defaultModel}}')
-    expect(componentWxml).toContain('issue553-update-abc')
-    expect(componentWxml).toContain('issue553-update-model')
-    expect(componentWxml).toContain('bindtap="__weapp_vite_inline"')
-    expect(componentWxml).toContain('data-wi-tap=')
-    expect(componentJs).toContain('updateAbc')
-    expect(componentJs).toContain('updateModelValue')
-  })
-
-  it('issue #554: injects slot metadata for v-for component default slots', async () => {
-    await runBuild()
-
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-554/index.json')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-554/index.wxml')
-    const pageJson = await fs.readJson(pageJsonPath) as { usingComponents?: Record<string, string> }
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-
-    expect(pageJson.usingComponents).toMatchObject({
-      'loop-slot-cell': '/components/issue-554/LoopSlotCell/index',
-    })
-    expect(pageWxml).toContain('loop-slot-cell')
-    expect(pageWxml).toContain('wx:for="{{items}}"')
-    expect(pageWxml).toContain('wx:for-item="__wv_item_')
-    expect(pageWxml).toContain('wx:key="key"')
-    expect(pageWxml).toContain(`vue-slots="{{__wv_bind_0[__wv_index_1]}}"`)
-    expect(pageWxml).toContain('<image class="issue554-image" src="{{__wv_item_0.src}}" mode="aspectFit" />')
-  })
-
-  it('issue #615: compiles scoped slot v-for owner list through a safe runtime binding', async () => {
-    await runIssue615AugmentedBuild()
-
-    const files = await scanFiles(DIST_ROOT)
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-615/index.wxml')
-    const scopedSlotWxmlFiles = files
-      .filter(file => file.startsWith('pages/issue-615/index.__scoped-slot-') && file.endsWith('.wxml'))
-      .sort()
-    const scopedSlotJsPath = path.join(DIST_ROOT, 'pages/issue-615/index.__scoped-slot-default-0.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const scopedSlotWxml = (await Promise.all(scopedSlotWxmlFiles.map(file => fs.readFile(path.join(DIST_ROOT, file), 'utf-8')))).join('\n')
-    const scopedSlotJs = await fs.readFile(scopedSlotJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('generic:scoped-slots-default=')
-    expect(scopedSlotWxmlFiles).toContain('pages/issue-615/index.__scoped-slot-default-0.wxml')
-    expect(scopedSlotWxml).toContain('tabbar-item wx:for="{{__wv_bind_0}}"')
-    expect(scopedSlotWxml).not.toContain('wx:for="{{__wvOwner.list}}"')
-    expect(scopedSlotWxml).toContain('data-issue615-label="{{item.label}}"')
-    expect(scopedSlotWxml).toContain('{{item.label}}')
-    expect(scopedSlotJs).toContain('__wvOwnerProxy')
-    expect(scopedSlotJs).toContain('list')
-    expect(scopedSlotJs).not.toContain('模板 v-for 数据源表达式执行失败: __wv_bind_0')
-    expect(scopedSlotJs).not.toContain('模板运行时表达式执行失败: __wv_bind_0 = list')
-  })
-
-  it('issue #829: keeps nested scoped-slot function props on the owner path', async () => {
-    await runAggregateBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-829/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-829/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf8')
-    const scopedSlotWxmlFiles = (await scanFiles(path.join(DIST_ROOT, 'pages/issue-829')))
-      .filter(file => file.startsWith('index.__scoped-slot-default-') && file.endsWith('.wxml'))
-    const scopedSlotWxml = await Promise.all(scopedSlotWxmlFiles.map(async file => await fs.readFile(path.join(DIST_ROOT, 'pages/issue-829', file), 'utf8')))
-
-    expect(pageWxml).toContain('query-fn="{{queryFn}}"')
-    expect(pageJs).toContain('__wevuFunctionPropPaths')
-    expect(pageJs).toContain('"queryFn"')
-    expect(scopedSlotWxml.some(content => content.includes('query-fn="{{__wvOwner.queryFn}}"'))).toBe(true)
-  })
-
-  it('issue #621: compiles inline assignment events against setup ref values', async () => {
-    await runIssue621AugmentedBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-621/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-621/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-621 inline assignment event')
-    expect(pageWxml).toContain('data-wi-tap="i0"')
-    expect(pageWxml).toContain('data-wi-tap="i1"')
-    expect(pageWxml).toContain('bindtap="__weapp_vite_inline"')
-    expect(pageJs).toContain('__weapp_vite_inline_map')
-    expect(pageJs).toContain('ctx.count.value += 1')
-    expect(pageJs).toContain('ctx.explicitCount.value += 1')
-    expect(pageJs).toContain('ctx.derivedCount.value = ctx.derivedCount.value + 1')
-    expect(pageJs).toContain('++ctx.prefixCount.value')
-    expect(pageJs).toContain('ctx.conditionalCount.value > 0 ? ctx.conditionalCount.value = ctx.conditionalCount.value + 2 : ctx.conditionalCount.value = ctx.conditionalCount.value + 1')
-    expect(pageJs).toContain('ctx.sequenceCount.value = ctx.sequenceCount.value + 1, ctx.sequenceCount.value++')
-    expect(pageJs).toContain('ctx.assignArgument(ctx.argumentCount.value)')
-    expect(pageJs).toContain('ctx.assignShorthand({ shorthandCount: ctx.shorthandCount.value })')
-    expect(pageJs).toContain('ctx.nestedState.count.value += 1')
-    expect(pageJs).not.toContain('ReferenceError')
-    expect(pageJs).not.toContain('ctx.explicitCount.value.value')
-    expect(pageJs).not.toContain('ctx.nestedState.count.value.value')
-  })
-
-  it('issue #642: keeps slot bridge props available to performance setData pick', async () => {
-    await runIssue642Build()
-
-    const pageWxmlPath = path.join(ISSUE_642_DIST_ROOT, 'pages/issue-642/index.wxml')
-    const pageJsPath = path.join(ISSUE_642_DIST_ROOT, 'pages/issue-642/index.js')
-    const scopedSlotWxmlPath = path.join(ISSUE_642_DIST_ROOT, 'pages/issue-642/index.__scoped-slot-default-0.wxml')
-    const slotProbeWxmlPath = path.join(ISSUE_642_DIST_ROOT, 'components/issue-642/SlotProbe/index.wxml')
-    const slotProbeJsPath = path.join(ISSUE_642_DIST_ROOT, 'components/issue-642/SlotProbe/index.js')
-    const scopedProbeWxmlPath = path.join(ISSUE_642_DIST_ROOT, 'components/issue-642/ScopedSlotProbe/index.wxml')
-    const scopedProbeJsPath = path.join(ISSUE_642_DIST_ROOT, 'components/issue-642/ScopedSlotProbe/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const scopedSlotWxml = await fs.readFile(scopedSlotWxmlPath, 'utf-8')
-    const slotProbeWxml = await fs.readFile(slotProbeWxmlPath, 'utf-8')
-    const slotProbeJs = await fs.readFile(slotProbeJsPath, 'utf-8')
-    const scopedProbeWxml = await fs.readFile(scopedProbeWxmlPath, 'utf-8')
-    const scopedProbeJs = await fs.readFile(scopedProbeJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue642-slot-probe')
-    expect(pageWxml).toContain('vue-slots="{{ {header:true,default:true} }}"')
-    expect(pageWxml).toContain('vue-slots="{{ {default:true} }}"')
-    expect(pageWxml).not.toContain('vue-slots="{{{')
-    expect(pageWxml).not.toContain('vue-slots="{{ {[')
-    expect(pageWxml).toContain('generic:scoped-slots-default=')
-    expectSetDataPickKeys(pageJs, ['vueSlots', '__wvSlotOwnerId', '__wvSlotScope'])
-    expectSetDataPickOmitsKeys(pageJs, ['__wv_bind_0', '__wv_bind_839'])
-    expect(scopedSlotWxml).toContain('data-issue642-slot-state="scoped-provided"')
-    expect(scopedSlotWxml).toContain('data-issue642-scoped-value="{{__wvSlotPropsData.io}}"')
-    expect(scopedSlotWxml).toContain('{{__wvSlotPropsData.io}}')
-    expect(slotProbeWxml).toContain(`<block wx:if="{{vueSlots&&vueSlots.header}}">`)
-    expect(slotProbeWxml).toContain(`<block wx:if="{{vueSlots&&vueSlots.default}}">`)
-    expect(slotProbeJs).toContain('"vueSlots"')
-    expectSetDataPickKeys(slotProbeJs, ['vueSlots', '__wvSlotOwnerId', '__wvSlotScope'])
-    expect(scopedProbeWxml).toContain('<scoped-slots-default wx:if="{{__wvSlotOwnerId}}"')
-    expect(scopedProbeWxml).toContain(`__wvSlotProps="{{['io',1234]}}"`)
-    expect(scopedProbeJs).toContain('"__wvSlotOwnerId"')
-    expect(scopedProbeJs).toContain('"__wvSlotScope"')
-    expectSetDataPickKeys(scopedProbeJs, ['vueSlots', '__wvSlotOwnerId', '__wvSlotScope'])
-  })
-
-  it('issue #674: injects vueSlots into slot-aware components without declared props', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-674/index.wxml')
-    const slotProbeWxmlPath = path.join(DIST_ROOT, 'components/issue-674/SlotProbe/index.wxml')
-    const slotProbeJsPath = path.join(DIST_ROOT, 'components/issue-674/SlotProbe/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const slotProbeWxml = await fs.readFile(slotProbeWxmlPath, 'utf-8')
-    const slotProbeJs = await fs.readFile(slotProbeJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue674-slot-probe')
-    expect(pageWxml).toContain('vue-slots=')
-    expect(pageWxml).toContain('header:true')
-    expect(pageWxml).toContain('default:true')
-    expect(slotProbeWxml).toContain('<slot />')
-    expect(slotProbeWxml).toContain('<slot name="header" />')
-    expect(slotProbeWxml).not.toContain('vueSlots&&vueSlots')
-    expect(slotProbeJs).toContain('"vueSlots"')
-    expect(slotProbeJs).toContain('"__wvSlotOwnerId"')
-    expect(slotProbeJs).toContain('"__wvSlotScope"')
-  })
-
-  it('issue #424: avoids duplicated output for imported src/assets images', async () => {
-    await runBuild()
-
-    const files = await scanFiles(DIST_ROOT)
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-424/index.js')
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-424/index.wxml')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-
-    expect(files).not.toContain('assets/images/home/goods-1.png')
-    expect(files).not.toContain('assets/images/home/banner-1.jpg')
-    expect(files.some(file => /^goods-1-[a-z0-9]+\.png$/.test(file))).toBe(true)
-    expect(files.some(file => /^banner-1-[a-z0-9]+\.jpg$/.test(file))).toBe(true)
-    expect(issuePageWxml).toContain('issue-424 duplicated imported src asset output')
-    expect(issuePageJs).toContain('_runE2E')
-    expect(issuePageJs).toContain('goods-1-')
-    expect(issuePageJs).toContain('banner-1-')
-  })
-
-  it('issue #420: injects WebSocket globals for socket.io-client entry pages', async () => {
-    await runBuild()
-
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-420/index.js')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-420/index.wxml')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-420 socket.io-client bootstrap')
-    expect(pageJs).toContain('transportName')
-    expect(pageJs).toContain('socket.invalid/github-issues')
-    expect(pageJs).toContain('/* __wvRGC__ */')
-    expect(pageJs).toContain('installWebRuntimeGlobals({ targets: [')
-    expect(pageJs).toContain('"WebSocket"')
-    expect(pageJs).toContain('|| globalThis')
-    expect(pageJs).toContain('var WebSocket =')
-    expect(pageJs).toContain('.WebSocket')
-    expect(pageJs).toContain('var URL =')
-    expect(pageJs).toContain('.URL')
-  })
-
-  it('issue #448: injects the next batch of web runtime globals on demand', async () => {
-    await runBuild()
-
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-448/index.js')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-448/index.wxml')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-448 next web runtime globals')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('microtaskState')
-    expect(pageJs).toContain('installWebRuntimeGlobals({ targets: [')
-    expect(pageJs).toContain('"queueMicrotask"')
-    expect(pageJs).toContain('const encoded = btoa("AB");')
-    expect(pageJs).toContain('URL.parse("123", "fake://abc")')
-    expect(pageJs).toContain('URL.canParse(')
-    expect(pageJs).toContain('searchParams.sort()')
-    expect(pageJs).toContain('headers.getSetCookie()')
-    expect(pageJs).toContain('Response.json(')
-    expect(pageJs).toContain('Response.error()')
-    expect(pageJs).toContain('function _runE2E() {')
-    expect(pageJs).toContain('function _runFormDataUploadE2E()')
-    expect(pageJs).toContain('globalThis.wx')
-    expect(pageJs).toContain('.downloadFile({')
-    expect(pageJs).toContain('new Blob([')
-    expect(pageJs).toContain('new File([')
-    expect(pageJs).toContain('FormData')
-  })
-
-  it('issue #457: keeps injected web runtime code readable in production output', async () => {
-    await runBuild()
-
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-457/index.js')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-457/index.wxml')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const lineCount = pageJs.split('\n').length
-
-    expect(pageWxml).toContain('issue-457 web runtime debug readability')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('installWebRuntimeGlobals({ targets: [')
-    expect(pageJs).toContain('"atob"')
-    expect(pageJs).toContain('"btoa"')
-    expect(pageJs).toContain('const encoded = btoa("issue-457");')
-    expect(lineCount).toBeGreaterThan(5)
-  })
-
-  it('issue #475: keeps emitted github-issues js traceable back to vue page and component sources', async () => {
-    await runBuildWithSourcemap()
-
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-475/index.js')
-    const componentJsPath = path.join(DIST_ROOT, 'components/issue-475/SourceMapProbe/index.js')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-475/index.wxml')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const componentJs = await fs.readFile(componentJsPath, 'utf-8')
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-
-    expect(pageWxml).toContain('class="issue475-page"')
-    expect(pageWxml).toContain('{{pageLabel}}')
-    expect(pageWxml).toContain('{{pageTraceLabel}}')
-    expect(pageWxml).toContain('{{pageEnvLabel}}')
-    expect(pageWxml).toContain('<source-map-probe />')
-    expect(pageJs).toContain('//#region src/pages/issue-475/index.vue')
-    expect(pageJs).toContain('issue-475 page marker')
-    expect(pageJs).toMatch(createCachedEnvLinePattern('pageEnv'))
-    expect(pageJs).not.toContain('const pageEnv = {\n')
-    expect(componentJs).toContain('//#region src/components/issue-475/SourceMapProbe/index.vue')
-    expect(componentJs).toContain('issue-475 component marker')
-    expect(componentJs).toMatch(createCachedEnvLinePattern('componentEnv'))
-    expect(componentJs).not.toContain('const componentEnv = {\n')
-  })
-
-  it('issue #769: keeps native ts sourcemaps aligned after local npm rewrite', async () => {
-    await runBuildWithSourcemap()
-
-    const sourceFile = 'src/subpackages/item/issue-769/index.ts'
-    const pageJsPath = path.join(DIST_ROOT, 'subpackages/item/issue-769/index.js')
-    const pageMapPath = `${pageJsPath}.map`
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageMap = JSON.parse(await fs.readFile(pageMapPath, 'utf-8')) as Parameters<typeof TraceMap>[0]
-    const sourceTs = await fs.readFile(path.join(APP_ROOT, sourceFile), 'utf-8')
-    const sourceLines = sourceTs.split('\n')
-    const sourceMap = new TraceMap(pageMap)
-    const mappedMarkers = [
-      ['issue 769 sourcemap marker', `camelCase('issue 769 sourcemap marker')`],
-      ['Page({', 'Page({'],
-      ['_runE2E() {', '_runE2E() {'],
-      ['issue769SourcemapMarker', `ok: issue769NpmMarker === 'issue769SourcemapMarker'`],
-    ]
-
-    expect(pageJs).toMatch(/require\((['"])\.\.\/miniprogram_npm\/camelcase(?:\/index)?\1\)/)
-    expect(await fs.pathExists(pageMapPath)).toBe(true)
-    for (const [generatedMarker, sourceMarker] of mappedMarkers) {
-      const expectedSourceLine = sourceLines.findIndex(line => line.includes(sourceMarker)) + 1
-      const originalPosition = originalPositionFor(sourceMap, findGeneratedPosition(pageJs, generatedMarker))
-      expect(expectedSourceLine).toBeGreaterThan(0)
-      expect(originalPosition.source?.replaceAll('\\', '/')).toContain(sourceFile)
-      expect(originalPosition.line).toBe(expectedSourceLine)
-    }
-  })
-
-  it('issue #479: injects indirect wevu page feature hooks from local helpers', async () => {
-    await runBuild()
-
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-479/index.js')
-    const issuePageJsonPath = path.join(DIST_ROOT, 'pages/issue-479/index.json')
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-479/index.wxml')
-
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-    const issuePageJson = await fs.readFile(issuePageJsonPath, 'utf-8')
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-479 indirect page feature hooks')
-    expect(issuePageJs).toContain('useIssue479PageFeatureHooks')
-    expect(issuePageJs).toContain('enableOnPullDownRefresh: true')
-    expect(issuePageJs).toContain('enableOnReachBottom: true')
-    expect(issuePageJson).toContain('"enablePullDownRefresh": true')
-    expect(issuePageJson).toContain('"onReachBottomDistance": 50')
-    expect(issuePageJs).not.toContain('pages/issue-479/usePageFeatureHooks')
-  })
-
-  it('issue #695: keeps direct pull-down hook feature enabled for Vue SFC pages', async () => {
-    await runBuild()
-
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-695/index.js')
-    const issuePageJsonPath = path.join(DIST_ROOT, 'pages/issue-695/index.json')
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-695/index.wxml')
-
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-    const issuePageJson = await fs.readFile(issuePageJsonPath, 'utf-8')
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-695 direct pull-down hook')
-    expect(issuePageJs).toContain('enableOnPullDownRefresh: true')
-    expect(issuePageJson).toContain('"enablePullDownRefresh": true')
-  })
-
-  it('issue #484: keeps config.define import.meta.env member overrides in weapp-vite preprocessed files', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-484/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-484/index.js')
-    const helperJsPath = path.join(DIST_ROOT, 'pages/issue-484/define.js')
-    const issuePageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const helperJs = await fs.pathExists(helperJsPath)
-      ? await fs.readFile(helperJsPath, 'utf-8')
-      : pageJs
-
-    expect(issuePageWxml).toContain('issue-484 import.meta.env define override')
-    expect(pageJs).toContain('var pageDefinedFlag = 123456;')
-    expect(pageJs).toContain('const pageSummary = `issue-484 page define:')
-    expect(pageJs).toMatch(createSerializedJsonEntryPattern('MODE', 'production'))
-    expect(pageJs).not.toContain('ISSUE_484_FLAG')
-    expect(helperJs).toContain('var helperDefinedFlag = 123456;')
-    expect(helperJs).toContain('helperSummary = `issue-484 helper define:')
-    expect(helperJs).toMatch(createSerializedJsonEntryPattern('MODE', 'production'))
-    expect(helperJs).not.toContain('ISSUE_484_FLAG')
-  })
-
-  it('experiment: preserves block slot syntax in native github-issues fixtures', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/block-slot/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/block-slot/index.js')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/block-slot/index.json')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/block-slot-host/index.wxml')
-    const componentJsPath = path.join(DIST_ROOT, 'components/block-slot-host/index.js')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageJson = await fs.readJson(pageJsonPath)
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const componentJs = await fs.readFile(componentJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('<block slot="header">')
-    expect(pageWxml).toContain('header slot via block: {{headerLabel}}')
-    expect(pageWxml).toContain('default slot via block: {{bodyLabel}}')
-    expect(pageJs).toContain('toggleLabels')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJson.usingComponents).toMatchObject({
-      'block-slot-host': '../../components/block-slot-host/index',
-    })
-    expect(componentWxml).toContain('<slot name="header"></slot>')
-    expect(componentWxml).toContain('<slot></slot>')
-    expect(componentJs).toContain('multipleSlots: true')
-  })
-
-  it('issue #494: removes synthetic view wrappers from plain template v-slot content when possible', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-494/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-494/index.js')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-494/SlotHost/index.wxml')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('slot="icon"')
-    expect(pageWxml).toContain('data-probe="single-image"')
-    expect(pageWxml).toContain('src="{{iconSrc}}"')
-    expect(pageWxml).toContain('<weapp-slot-wrapper slot="header"><view class="issue494-header-probe"')
-    expect(pageWxml).toContain('header via template slot: {{headerLabel}}')
-    expect(pageWxml).toContain('default via template slot: {{bodyLabel}}')
-    expect(pageWxml).not.toContain('<view slot="icon">')
-    expect(pageWxml).not.toContain('<block slot="header">')
-    expect(pageJs).toContain('toggleLabels')
-    expect(pageJs).toContain('_runE2E')
-    expect(componentWxml).toContain('<slot name="icon" />')
-    expect(componentWxml).toContain('<slot name="header" />')
-    expect(componentWxml).toContain('<slot />')
-  })
-
-  it('issue #555: keeps slot attribute on v-if single root content instead of generated block', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-555/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-555/index.js')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-555/SlotCell/index.wxml')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const conditionalTextSlotStart = pageWxml.indexOf('<block wx:if="{{value}}"><text')
-    const conditionalTextSlotEnd = pageWxml.indexOf('{{value}}</text></block>', conditionalTextSlotStart)
-    const conditionalTextSlot = pageWxml.slice(conditionalTextSlotStart, conditionalTextSlotEnd)
-
-    expect(conditionalTextSlotStart).toBeGreaterThanOrEqual(0)
-    expect(conditionalTextSlotEnd).toBeGreaterThan(conditionalTextSlotStart)
-    expect(conditionalTextSlot).toContain('slot="text"')
-    expect(pageWxml).toContain('data-probe="conditional-text"')
-    expect(pageWxml).toContain('{{value}}</text></block>')
-    expect(pageWxml).not.toContain('<block slot="text"')
-    expect(pageWxml).not.toContain('<view slot="text">')
-    expect(pageJs).toContain('toggleValue')
-    expect(pageJs).toContain('_runE2E')
-    expect(componentWxml).toContain('<slot name="text" />')
-  })
-
-  it('issue #574: preserves template slot output order around implicit default children', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-574/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-574/index.js')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-574/VanCell/index.wxml')
-
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const defaultIndex = pageWxml.indexOf('data-order="default"')
-    const footer1Index = pageWxml.indexOf('data-order="footer1"')
-    const footer2Index = pageWxml.indexOf('data-order="footer2"')
-
-    expect(defaultIndex).toBeGreaterThanOrEqual(0)
-    expect(footer1Index).toBeGreaterThan(defaultIndex)
-    expect(footer2Index).toBeGreaterThan(footer1Index)
-    expect(pageWxml).toContain('van-cell vue-slots="{{ {footer1:true,footer2:true,default:true} }}"')
-    expect(pageWxml).not.toContain('vue-slots="{{{')
-    expect(pageWxml).not.toContain('vue-slots="{{ {[')
-    expect(pageWxml).toContain('<view slot="footer1" class="issue574-footer1" data-order="footer1" />')
-    expect(pageWxml).toContain('<text slot="footer2" class="issue574-footer2" data-order="footer2" />')
-    expect(pageJs).toContain('_runE2E')
-    expect(componentWxml).toContain('<slot />')
-    expect(componentWxml).toContain('<slot name="footer1" />')
-    expect(componentWxml).toContain('<slot name="footer2" />')
-  })
-
-  it('issue #581: compiles setup object reactive array and loading refs', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-581/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-581/index.js')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-581 reactive array flush')
-    expect(pageWxml).toContain('wx:for="{{back.state}}"')
-    expect(pageWxml).toContain('data-issue581-name="{{value.name}}"')
-    expect(pageWxml).toContain('data-loading="{{back.loading ? \'true\' : \'false\'}}"')
-    expect(pageWxml).toContain('data-row-count="{{back.state.length}}"')
-    expect(pageJs).toContain('fetchIssue581Rows')
-    expect(pageJs).toContain('_runE2E')
-  })
-
-  it('issue #590: compiles props-based class/style bindings for components without setup', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-590/index.wxml')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-590/index.json')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-590/SimpleBadge/index.wxml')
-    const componentJsPath = path.join(DIST_ROOT, 'components/issue-590/SimpleBadge/index.js')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJson = await fs.readJSON(pageJsonPath) as {
-      usingComponents?: Record<string, string>
-    }
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const componentJs = await fs.readFile(componentJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-590 no setup props computed style')
-    expect(pageJson.usingComponents?.['simple-badge']).toBe('/components/issue-590/SimpleBadge/index')
-    expect(componentWxml).toContain('id="issue590-badge"')
-    expect(componentWxml).toMatch(/class="\{\{__wv_cls_\d+\}\}"/)
-    expect(componentWxml).toMatch(/style="\{\{__wv_style_\d+\}\}"/)
-    expect(componentJs).toMatch(/__wv_cls_\d+/)
-    expect(componentJs).toMatch(/__wv_style_\d+/)
-    expect(componentJs).toContain('__wevuProps')
-    expect(componentJs).toContain('issue590-badge')
-    expect(componentJs).toContain('issue-590 badge')
-  })
-
-  it('issue #599: prefers props named data in generated computed bindings', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-599/index.wxml')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-599/index.json')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-599/DataPropProbe/index.wxml')
-    const componentJsPath = path.join(DIST_ROOT, 'components/issue-599/DataPropProbe/index.js')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJson = await fs.readJSON(pageJsonPath) as {
-      usingComponents?: Record<string, string>
-    }
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const componentJs = await fs.readFile(componentJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-599 data prop computed')
-    expect(pageJson.usingComponents?.['data-prop-probe']).toBe('/components/issue-599/DataPropProbe/index')
-    expect(componentWxml).toContain('data-issue599-label="{{data.label}}"')
-    expect(componentWxml).toMatch(/style="\{\{__wv_style_\d+\}\}"/)
-    expect(componentJs).toContain('allowNullPropInput: true')
-    expect(componentJs).toContain('this.data')
-    expect(componentJs).toContain('props: { data: {')
-    expect(componentJs).toContain('__wevuPropsDerivedKeys: ["data"]')
-  })
-
-  it('issue #627: compiles class/style defineProps probe', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-627/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-627/index.js')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-627/index.json')
-    const nativePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-627-native/index.wxml')
-    const nativePageJsPath = path.join(DIST_ROOT, 'pages/issue-627-native/index.js')
-    const nativePageJsonPath = path.join(DIST_ROOT, 'pages/issue-627-native/index.json')
-    const sfcHostPageWxmlPath = path.join(DIST_ROOT, 'pages/issue-627-sfc-host/index.wxml')
-    const sfcHostPageJsPath = path.join(DIST_ROOT, 'pages/issue-627-sfc-host/index.js')
-    const sfcHostPageJsonPath = path.join(DIST_ROOT, 'pages/issue-627-sfc-host/index.json')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-627/ReservedPropsProbe/index.wxml')
-    const componentJsPath = path.join(DIST_ROOT, 'components/issue-627/ReservedPropsProbe/index.js')
-    const nativeComponentWxmlPath = path.join(DIST_ROOT, 'components/issue-627/native-props-probe/index.wxml')
-    const nativeComponentJsPath = path.join(DIST_ROOT, 'components/issue-627/native-props-probe/index.js')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageJson = await fs.readJSON(pageJsonPath) as {
-      usingComponents?: Record<string, string>
-    }
-    const nativePageWxml = await fs.readFile(nativePageWxmlPath, 'utf-8')
-    const nativePageJs = await fs.readFile(nativePageJsPath, 'utf-8')
-    const nativePageJson = await fs.readJSON(nativePageJsonPath) as {
-      usingComponents?: Record<string, string>
-    }
-    const sfcHostPageWxml = await fs.readFile(sfcHostPageWxmlPath, 'utf-8')
-    const sfcHostPageJs = await fs.readFile(sfcHostPageJsPath, 'utf-8')
-    const sfcHostPageJson = await fs.readJSON(sfcHostPageJsonPath) as {
-      usingComponents?: Record<string, string>
-    }
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const componentJs = await fs.readFile(componentJsPath, 'utf-8')
-    const nativeComponentWxml = await fs.readFile(nativeComponentWxmlPath, 'utf-8')
-    const nativeComponentJs = await fs.readFile(nativeComponentJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-627 reserved props')
-    expect(pageJson.usingComponents?.['reserved-props-probe']).toBe('/components/issue-627/ReservedPropsProbe/index')
-    expect(pageJs).toContain('readProbe("#issue627-sfc-probe-literal")')
-    expect(pageJs).toContain('readProbe("#issue627-sfc-probe-dynamic")')
-    expect(pageWxml).toContain('class="issue-627-class-prop"')
-    expect(pageWxml).toContain('style="color: rgb(22 119 255);"')
-    expect(pageWxml).toContain('id="issue627-sfc-probe-literal"')
-    expect(pageWxml).toContain('id="issue627-sfc-probe-dynamic"')
-    expect(componentWxml).toContain('data-issue627-class="{{props.class}}"')
-    expect(componentWxml).toContain('data-issue627-style="{{props.style}}"')
-    expect(componentJs).toContain('props: {')
-    expect(componentJs).toContain('class: {')
-    expect(componentJs).toContain('style: {')
-    expect(componentJs).toContain('expose({ _runE2E })')
-
-    expect(nativePageJs).toContain('issue-627 native reserved props')
-    expect(nativePageWxml).toContain('issue627-native-title')
-    expect(nativePageWxml).toContain('class="issue-627-native-class-prop"')
-    expect(nativePageWxml).toContain('style="color: rgb(22, 119, 255);"')
-    expect(nativePageJs).toContain('selectComponent?.("#issue627-native-probe")')
-    expect(nativePageJson.usingComponents?.['native-props-probe']).toBe('/components/issue-627/native-props-probe/index')
-    expect(nativeComponentWxml).toContain('data-issue627-native-class="{{class}}"')
-    expect(nativeComponentWxml).toContain('data-issue627-native-style="{{style}}"')
-    expect(nativeComponentJs).toContain('"class": String')
-    expect(nativeComponentJs).toContain('"style": String')
-    expect(nativeComponentJs).toContain('"hidden": Boolean')
-    expect(nativeComponentJs).toContain('"dataFoo": String')
-    expect(nativeComponentJs).toContain('"data-foo": String')
-    expect(nativeComponentJs).toContain('"markFoo": String')
-    expect(nativeComponentJs).toContain('"mark:foo": String')
-    expect(nativeComponentJs).toContain('"slot": String')
-
-    expect(sfcHostPageJs).toContain('issue-627 sfc host reserved props')
-    expect(sfcHostPageJs).toContain('selectComponent?.("#issue627-sfc-host-literal")')
-    expect(sfcHostPageJson.usingComponents?.ReservedPropsProbe).toBe('/components/issue-627/ReservedPropsProbe/index')
-    expect(sfcHostPageWxml).toContain('class="issue-627-sfc-host-class-prop"')
-    expect(sfcHostPageWxml).toContain('style="color: rgb(22, 119, 255);"')
-    expect(sfcHostPageWxml).toContain('class="issue-627-sfc-host-dynamic-class"')
-    expect(sfcHostPageWxml).toContain('style="font-size: 32rpx;"')
-  })
-
-  it('issue #597: preserves same-name conditional slot branches', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-597/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-597/index.js')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-597/Issue597Card/index.wxml')
-
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue597-card vue-slots="{{ {header:true} }}"')
-    expect(pageWxml).not.toContain('vue-slots="{{{')
-    expect(pageWxml).not.toContain('vue-slots="{{ {[')
-    expect(pageWxml).toContain('<block wx:if="{{abc}}"><view slot="header" class="issue597-header-a" data-issue597-branch="if" /></block>')
-    expect(pageWxml).toContain('<block wx:else><text slot="header" class="issue597-header-b" data-issue597-branch="else" /></block>')
-    expect(pageWxml).not.toContain('<text slot="header" class="issue597-header-b" data-issue597-branch="else" />\n</issue597-card>')
-    expect(pageJs).toContain('_runE2E')
-    expect(componentWxml).toContain('<slot name="header" />')
-  })
-
-  it('issue #613: wraps forwarded slot outlets inside plain named slot content', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-613/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-613/index.js')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-613/Issue613Card/index.wxml')
-    const forwarderWxmlPath = path.join(DIST_ROOT, 'components/issue-613/Issue613ViewForwarder/index.wxml')
-    const forwarderJsonPath = path.join(DIST_ROOT, 'components/issue-613/Issue613ViewForwarder/index.json')
-    const appJsonPath = path.join(DIST_ROOT, 'app.json')
-    const wrapperWxmlPath = path.join(DIST_ROOT, 'weapp_vite_internal/slot-wrapper/index.wxml')
-    const wrapperJsonPath = path.join(DIST_ROOT, 'weapp_vite_internal/slot-wrapper/index.json')
-    const wrapperJsPath = path.join(DIST_ROOT, 'weapp_vite_internal/slot-wrapper/index.js')
-    const legacyForwarderWxmlPath = path.join(DIST_ROOT, 'components/issue-613/Issue613LegacyViewForwarder/index.wxml')
-    const legacyForwarderJsonPath = path.join(DIST_ROOT, 'components/issue-613/Issue613LegacyViewForwarder/index.json')
-    const blockForwarderWxmlPath = path.join(DIST_ROOT, 'components/issue-613/block-forwarder/index.wxml')
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const forwarderWxml = await fs.readFile(forwarderWxmlPath, 'utf-8')
-    const appJson = await fs.readJSON(appJsonPath) as { usingComponents?: Record<string, string> }
-    const forwarderJson = await fs.readJSON(forwarderJsonPath) as { usingComponents?: Record<string, string> }
-    const wrapperWxml = await fs.readFile(wrapperWxmlPath, 'utf-8')
-    const wrapperJson = await fs.readJSON(wrapperJsonPath) as { component?: boolean }
-    const wrapperJs = await fs.readFile(wrapperJsPath, 'utf-8')
-    const legacyForwarderWxml = await fs.readFile(legacyForwarderWxmlPath, 'utf-8')
-    const legacyForwarderJson = await fs.readJSON(legacyForwarderJsonPath) as { usingComponents?: Record<string, string> }
-    const blockForwarderWxml = await fs.readFile(blockForwarderWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('<issue613-card data-issue613-case="compiled-fragment"')
-    expect(pageWxml).toContain('<weapp-slot-wrapper slot="header"><slot /></weapp-slot-wrapper>')
-    expect(pageWxml).not.toContain('<slot slot="header"')
-    expect(pageWxml).toContain('issue613-block-forwarder')
-    expect(forwarderWxml).toContain('<issue613-card vue-slots="{{ {header:true,footer:true} }}"')
-    expect(forwarderWxml).not.toContain('vue-slots="{{{')
-    expect(forwarderWxml).not.toContain('vue-slots="{{ {[')
-    expect(forwarderWxml).toContain('<weapp-slot-wrapper slot="header"><slot /></weapp-slot-wrapper>')
-    expect(forwarderWxml).toContain('<weapp-slot-wrapper slot="footer"><slot name="footer" /></weapp-slot-wrapper>')
-    expect(appJson.usingComponents?.['weapp-slot-wrapper']).toBeUndefined()
-    expect(forwarderJson.usingComponents?.['weapp-slot-wrapper']).toBe('/weapp_vite_internal/slot-wrapper/index')
-    expect(wrapperWxml).toBe('<slot></slot>')
-    expect(wrapperJson.component).toBe(true)
-    expect(wrapperJs).toContain('virtualHost:true')
-    expect(wrapperJs).toContain('multipleSlots:true')
-    expect(legacyForwarderWxml).toContain('<view slot="header"><slot /></view>')
-    expect(legacyForwarderJson.usingComponents?.['weapp-slot-wrapper']).toBeUndefined()
-    expect(forwarderWxml).not.toContain('<slot slot="header"')
-    expect(forwarderWxml).not.toContain('<slot slot="footer"')
-    expect(forwarderWxml).not.toContain('slot-wrapper=')
-    expect(forwarderWxml).not.toContain('slot-wrapper-footer=')
-    expect(blockForwarderWxml).toContain('<block slot="header">')
-    expect(blockForwarderWxml).toContain('<slot></slot>')
-    expect(pageJs).toContain('_runE2E')
-    expect(componentWxml).toContain('<slot name="header" />')
-    expect(componentWxml).toContain('<slot name="footer" />')
-  })
-
-  it('issue #595: builds scoped main package and selected subpackage only', async () => {
-    await runIssue595ScopedBuild()
-
-    const appJson = await fs.readJSON(path.join(ISSUE_595_DIST_ROOT, 'app.json')) as AppJsonWithSubPackages
-    const subPackageRoots = (appJson.subPackages ?? appJson.subpackages ?? []).map(entry => entry.root)
-    const files = await scanFiles(ISSUE_595_DIST_ROOT)
-
-    expect(appJson.pages).toContain('pages/block-slot/index')
-    expect(appJson.pages).toContain('pages/issue-590/index')
-    expect(subPackageRoots).toEqual(['subpackages/item'])
-    expect(files).toContain('pages/issue-590/index.js')
-    expect(files).toContain('subpackages/item/index.js')
-    expect(files).toContain('subpackages/item/login-required/index.js')
-    expect(files).not.toContain('subpackages/user/index.js')
-    expect(files).not.toContain('subpackages/issue-466/index.js')
-  })
-
-  it('issue #500: compiles missing inject continuation probe', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-500/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-500/index.js')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-500 inject missing key continuation')
-    expect(pageWxml).toContain('inject after line: {{continuationText}}')
-    expect(pageWxml).toContain('data-continuation="{{continuationText}}"')
-    expect(pageWxml).toContain('data-missing-type="{{missingType}}"')
-    expect(pageJs).toContain('issue-500:missing-token')
-    expect(pageJs).toContain('_runE2E')
-  })
-
-  it('issue #502: keeps structural directives on slot outlet elements', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-502/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-502/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-502 slot condition outlet')
-    expect(pageWxml).toContain('<block wx:if="{{abc}}"><scoped-slots-default wx:if="{{__wvSlotOwnerId}}"')
-    expect(pageWxml).toContain('<block wx:elif="{{efg}}"><scoped-slots-default wx:if="{{__wvSlotOwnerId}}"')
-    expect(pageWxml).toContain('<block wx:else><scoped-slots-default wx:if="{{__wvSlotOwnerId}}"')
-    expect(pageWxml).toContain('<block wx:else><slot /></block></block><block wx:elif="{{efg}}">')
-    expect(pageWxml).toContain('<block wx:else><slot /></block></block><block wx:else>')
-    expect(pageWxml).toContain('<block wx:else><slot /></block></block><view class="issue502-action"')
-    expect(pageWxml).not.toContain('<slot /><slot /><slot />')
-    expect(pageJs).toContain('toggleBranch')
-    expect(pageJs).toContain('_runE2E')
-  })
-
-  it('issue #521: keeps generated scoped slot components virtual for flex layouts', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-521/index.wxml')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-521/index.json')
-    const scopedSlotJsonPath = path.join(DIST_ROOT, 'pages/issue-521/index.__scoped-slot-default-0.json')
-    const scopedSlotWxmlPath = path.join(DIST_ROOT, 'pages/issue-521/index.__scoped-slot-default-0.wxml')
-    const hostWxmlPath = path.join(DIST_ROOT, 'components/issue-521/ScopedFlexHost/index.wxml')
-    const runtime = await findWevuRuntimeChunk(
-      DIST_ROOT,
-      code => code.includes('virtualHost') && /options:\s*\{\s*virtualHost:\s*(?:true|!0)\s*\}/.test(code),
-      'scoped slot virtualHost options',
-    )
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJson = await fs.readJson(pageJsonPath) as { usingComponents?: Record<string, string> }
-    const scopedSlotJson = await fs.readJson(scopedSlotJsonPath) as { component?: boolean, styleIsolation?: string }
-    const scopedSlotWxml = stripGeneratedScopeAttrs(await fs.readFile(scopedSlotWxmlPath, 'utf-8'))
-    const hostWxml = await fs.readFile(hostWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-521 scoped slot flex layout')
-    expect(pageWxml).toContain('generic:scoped-slots-default=')
-    expect(pageWxml).toContain(SLOT_OWNER_ATTR)
-    expect(pageJson.usingComponents).toMatchObject({
-      'scoped-flex-host': '/components/issue-521/ScopedFlexHost/index',
-      'flex-item': '/components/issue-521/FlexItem/index',
-    })
-    expect(Object.values(pageJson.usingComponents ?? {})).toContain('/pages/issue-521/index.__scoped-slot-default-0')
-    expect(scopedSlotJson).toMatchObject({
-      component: true,
-      styleIsolation: 'apply-shared',
-    })
-    expect(scopedSlotWxml).toContain('<flex-item label="A" value="{{__wvSlotPropsData.xyz}}" /><flex-item label="B" value="{{__wvSlotPropsData.xyz}}" />')
-    expect(await fs.readFile(scopedSlotWxmlPath.replace(/\.wxml$/, '.js'), 'utf-8')).toContain('createWevuScopedSlotComponent()')
-    expect(hostWxml).toContain('<scoped-slots-default wx:if="{{__wvSlotOwnerId}}"')
-    expect(runtime.code).toContain('virtualHost')
-    expect(runtime.code).toMatch(/options:\s*\{\s*virtualHost:\s*(?:true|!0)\s*\}/)
-  })
-
-  it('issue #510: augmented plain default slots preserve projected provide/inject scope', async () => {
-    await runIssue510AugmentedBuild()
-
-    const pageWxmlPath = path.join(ISSUE_510_DIST_ROOT, 'pages/issue-510/index.wxml')
-    const pageJsonPath = path.join(ISSUE_510_DIST_ROOT, 'pages/issue-510/index.json')
-    const scopedSlotWxmlPath = path.join(ISSUE_510_DIST_ROOT, 'pages/issue-510/index.__scoped-slot-default-0.wxml')
-    const hostWxmlPath = path.join(ISSUE_510_DIST_ROOT, 'components/issue-510/AugmentedSlotHost/index.wxml')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJson = await fs.readJson(pageJsonPath) as { usingComponents?: Record<string, string> }
-    const scopedSlotWxml = stripGeneratedScopeAttrs(await fs.readFile(scopedSlotWxmlPath, 'utf-8'))
-    const hostWxml = await fs.readFile(hostWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-510 augmented slot provide inject')
-    expect(pageWxml).toContain('generic:scoped-slots-default=')
-    expect(pageWxml).toContain(SLOT_OWNER_ATTR)
-    expect(pageWxml).not.toContain('<view class="issue510-wrapper"><augmented-slot-leaf /></view>')
-    expect(pageJson.usingComponents).toMatchObject({
-      'augmented-slot-host': '/components/issue-510/AugmentedSlotHost/index',
-      'augmented-slot-leaf': '/components/issue-510/AugmentedSlotLeaf/index',
-    })
-    expect(Object.values(pageJson.usingComponents ?? {})).toContain('/pages/issue-510/index.__scoped-slot-default-0')
-    expect(scopedSlotWxml).toContain('<view class="issue510-wrapper"><augmented-slot-leaf /></view>')
-    expect(hostWxml).toContain('<scoped-slots-default wx:if="{{__wvSlotOwnerId}}"')
-  })
-
-  it('issue #547: augmented nested default slots emit every scoped slot asset', async () => {
-    await runIssue547AugmentedBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-547/index.wxml')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-547/index.json')
-    const parentScopedSlotWxmlPath = path.join(DIST_ROOT, 'pages/issue-547/index.__scoped-slot-default-0.wxml')
-    const parentScopedSlotJsonPath = path.join(DIST_ROOT, 'pages/issue-547/index.__scoped-slot-default-0.json')
-    const childScopedSlotWxmlPath = path.join(DIST_ROOT, 'pages/issue-547/index.__scoped-slot-default-1.wxml')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJson = await fs.readJson(pageJsonPath) as { usingComponents?: Record<string, string> }
-    const parentScopedSlotWxml = stripGeneratedScopeAttrs(await fs.readFile(parentScopedSlotWxmlPath, 'utf-8'))
-    const parentScopedSlotJson = await fs.readJson(parentScopedSlotJsonPath) as { componentGenerics?: Record<string, true>, usingComponents?: Record<string, string> }
-    const childScopedSlotWxml = stripGeneratedScopeAttrs(await fs.readFile(childScopedSlotWxmlPath, 'utf-8'))
-
-    expect(pageWxml).toContain('issue-547 nested augmented slot')
-    expect(pageWxml).toContain('generic:scoped-slots-default=')
-    expect(pageWxml).not.toContain('<nested-slot-cell')
-    expect(pageJson.usingComponents).toMatchObject({
-      'nested-slot-group': '/components/issue-547/NestedSlotGroup/index',
-      'nested-slot-cell': '/components/issue-547/NestedSlotCell/index',
-      'nested-slot-image': '/components/issue-547/NestedSlotImage/index',
-    })
-    expect(Object.values(pageJson.usingComponents ?? {})).toContain('/pages/issue-547/index.__scoped-slot-default-0')
-    expect(parentScopedSlotWxml).toContain('<nested-slot-cell generic:scoped-slots-default=')
-    expect(parentScopedSlotWxml).not.toContain('<nested-slot-image')
-    expect(parentScopedSlotJson.componentGenerics).toBeUndefined()
-    expect(Object.values(parentScopedSlotJson.usingComponents ?? {})).toContain('/pages/issue-547/index.__scoped-slot-default-1')
-    expect(childScopedSlotWxml).toContain('<nested-slot-image />')
-  })
-
-  it('issue #558: augmented scoped slot runtime bindings read owner proxy', async () => {
-    await runIssue558AugmentedBuild()
-
-    const files = await scanFiles(DIST_ROOT)
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-558/index.wxml')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-558/index.json')
-    const cellWxmlPath = path.join(DIST_ROOT, 'components/issue-558/Cell/index.wxml')
-    const defaultScopedCellWxmlPath = path.join(DIST_ROOT, 'components/issue-558/DefaultScopedCell/index.wxml')
-    const listScopedCellWxmlPath = path.join(DIST_ROOT, 'components/issue-558/ListScopedCell/index.wxml')
-    const namedSlotCardWxmlPath = path.join(DIST_ROOT, 'components/issue-558/NamedSlotCard/index.wxml')
-    const namedSlotCardJsPath = path.join(DIST_ROOT, 'components/issue-558/NamedSlotCard/index.js')
-    const nestedSlotCellWxmlPath = path.join(DIST_ROOT, 'components/issue-558/Issue558NestedSlotCell/index.wxml')
-    const nestedSlotGroupWxmlPath = path.join(DIST_ROOT, 'components/issue-558/Issue558NestedSlotGroup/index.wxml')
-    const scopedSlotFiles = files
-      .filter(file => file.startsWith('pages/issue-558/index.__scoped-slot-') && file.endsWith('.wxml'))
-      .sort()
-    const scopedSlotJsFiles = files
-      .filter(file => file.startsWith('pages/issue-558/index.__scoped-slot-') && file.endsWith('.js'))
-      .sort()
-    const scopedSlotJsonFiles = files
-      .filter(file => file.startsWith('pages/issue-558/index.__scoped-slot-') && file.endsWith('.json'))
-      .sort()
-    const runtime = await findWevuRuntimeChunk(
-      DIST_ROOT,
-      code => code.includes('__wvOwnerProxy'),
-      'issue #558 owner proxy runtime',
-    )
-
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-    const pageJson = await fs.readJson(pageJsonPath) as { usingComponents?: Record<string, string> }
-    const scopedSlotWxml = stripGeneratedScopeAttrs((await Promise.all(scopedSlotFiles.map(file => fs.readFile(path.join(DIST_ROOT, file), 'utf-8')))).join('\n'))
-    const scopedSlotJs = (await Promise.all(scopedSlotJsFiles.map(file => fs.readFile(path.join(DIST_ROOT, file), 'utf-8')))).join('\n')
-    const scopedSlotJson = (await Promise.all(scopedSlotJsonFiles.map(file => fs.readFile(path.join(DIST_ROOT, file), 'utf-8')))).join('\n')
-    const cellWxml = await fs.readFile(cellWxmlPath, 'utf-8')
-    const defaultScopedCellWxml = await fs.readFile(defaultScopedCellWxmlPath, 'utf-8')
-    const listScopedCellWxml = await fs.readFile(listScopedCellWxmlPath, 'utf-8')
-    const namedSlotCardWxml = await fs.readFile(namedSlotCardWxmlPath, 'utf-8')
-    const namedSlotCardJs = await fs.readFile(namedSlotCardJsPath, 'utf-8')
-    const nestedSlotCellWxml = await fs.readFile(nestedSlotCellWxmlPath, 'utf-8')
-    const nestedSlotGroupWxml = await fs.readFile(nestedSlotGroupWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-558 augmented slot computed binding')
-    expect(pageWxml).toContain('generic:scoped-slots-default=')
-    expect(pageWxml).toContain('generic:scoped-slots-header=')
-    expect(pageWxml).toContain('generic:scoped-slots-footer=')
-    expect(pageJson.usingComponents).toMatchObject({
-      'cell': '/components/issue-558/Cell/index',
-      'default-scoped-cell': '/components/issue-558/DefaultScopedCell/index',
-      'issue558-nested-slot-cell': '/components/issue-558/Issue558NestedSlotCell/index',
-      'issue558-nested-slot-group': '/components/issue-558/Issue558NestedSlotGroup/index',
-      'issue-558-render-probe': '/components/issue-558/Issue558RenderProbe/index',
-      'list-scoped-cell': '/components/issue-558/ListScopedCell/index',
-      'named-slot-card': '/components/issue-558/NamedSlotCard/index',
-    })
-    expect(Object.values(pageJson.usingComponents ?? {}).some(value => value.startsWith('/pages/issue-558/index.__scoped-slot-'))).toBe(true)
-    expect(scopedSlotFiles.length).toBeGreaterThanOrEqual(5)
-    expect(scopedSlotJsFiles.length).toBe(scopedSlotFiles.length)
-    expect(scopedSlotJsonFiles.length).toBe(scopedSlotFiles.length)
-    expect(pageWxml).toMatch(/<issue-558-render-probe case-name="plainDefault" value="\{\{__wv_bind_\d+\}\}"/)
-    expect(scopedSlotWxml).not.toContain('case-name="plainDefault"')
-    expect(scopedSlotJson).toContain('/components/issue-558/Issue558RenderProbe/index')
-    for (const caseName of [
-      'namedHeader',
-      'explicitDefault',
-      'namedScopedFooter',
-      'defaultScoped',
-      'listScoped',
-      'nestedOuter',
-      'nestedDefault',
-    ]) {
-      expect(scopedSlotWxml).toMatch(new RegExp(`<issue-558-render-probe case-name="${caseName}" value="\\{\\{__wv_bind_\\d+\\}\\}"`))
-    }
-    expect(scopedSlotJs).toContain('__wvOwnerProxy')
-    expect(scopedSlotJs).toContain('.func')
-    expect(scopedSlotJs).toContain('.text')
-    expect(scopedSlotJs).toContain('.headerText')
-    expect(scopedSlotJs).toContain('.defaultText')
-    expect(scopedSlotJs).toContain('.nestedText')
-    expect(scopedSlotJs).toContain('__wvSlotPropsData')
-    expect(scopedSlotJs).toContain('.suffix')
-    expect(scopedSlotJs).toContain('.label')
-    expect(scopedSlotJs).toContain('.index')
-    expect(scopedSlotJs).not.toContain('__wvOwner.func')
-    expect(cellWxml).toContain('<slot />')
-    expect(cellWxml).not.toContain('<scoped-slots-default')
-    expect(defaultScopedCellWxml).toContain('__wvSlotProps="{{[\'label\',label,\'count\',count]}}"')
-    expect(listScopedCellWxml).toContain('wx:for="{{rows}}"')
-    expect(listScopedCellWxml).toContain('__wvSlotScope="{{__wvSlotScope}}"')
-    expect(namedSlotCardWxml).toContain('<slot name="header" />')
-    expect(namedSlotCardWxml).toContain('<scoped-slots-header')
-    expect(namedSlotCardWxml).toContain('<scoped-slots-default')
-    expect(namedSlotCardWxml).toContain('<scoped-slots-footer')
-    expect(namedSlotCardJs).toContain('properties')
-    expect(namedSlotCardJs).toContain('__wvSlotOwnerId')
-    expect(nestedSlotCellWxml).not.toContain('<scoped-slots-default')
-    expect(nestedSlotGroupWxml).toContain('<slot />')
-    expect(nestedSlotGroupWxml).toContain('<scoped-slots-default')
-    expect(runtime.code).toContain('__wvOwnerProxy')
-  })
-
-  it('issue #564: native component default content stays inline inside augmented scoped slot assets', async () => {
-    await runIssue564AugmentedBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-564/index.wxml')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-564/index.json')
-    const scopedSlotWxmlPath = path.join(DIST_ROOT, 'pages/issue-564/index.__scoped-slot-default-0.wxml')
-    const scopedSlotJsonPath = path.join(DIST_ROOT, 'pages/issue-564/index.__scoped-slot-default-0.json')
-    const nestedScopedSlotWxmlPath = path.join(DIST_ROOT, 'pages/issue-564/index.__scoped-slot-default-1.wxml')
-    const nativeTabbarWxmlPath = path.join(DIST_ROOT, 'components/issue-564/native-tabbar/index.wxml')
-    const nativeTabbarJsPath = path.join(DIST_ROOT, 'components/issue-564/native-tabbar/index.js')
-    const nativeTabbarJsonPath = path.join(DIST_ROOT, 'components/issue-564/native-tabbar/index.json')
-    const nativeTabbarGenericJsonPath = path.join(DIST_ROOT, 'components/issue-564/native-tabbar/__weapp_vite_scoped_slot_generic_component.json')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJson = await fs.readJson(pageJsonPath) as { usingComponents?: Record<string, string> }
-    const scopedSlotWxml = await fs.readFile(scopedSlotWxmlPath, 'utf-8')
-    const scopedSlotJson = await fs.readJson(scopedSlotJsonPath) as { usingComponents?: Record<string, string> }
-    const nativeTabbarWxml = await fs.readFile(nativeTabbarWxmlPath, 'utf-8')
-    const nativeTabbarJs = await fs.readFile(nativeTabbarJsPath, 'utf-8')
-    const nativeTabbarJson = await fs.readJson(nativeTabbarJsonPath) as { componentGenerics?: Record<string, { default?: string }> }
-    const nativeTabbarGenericJson = await fs.readJson(nativeTabbarGenericJsonPath) as { component?: boolean, options?: { virtualHost?: boolean } }
-
-    expect(pageWxml).toContain('issue-564 native nested scoped slot')
-    expect(pageWxml).toContain('generic:scoped-slots-default=')
-    expect(pageJson.usingComponents).toMatchObject({
-      'issue-564-native-tabbar': '/components/issue-564/native-tabbar/index',
-      'issue-564-native-tabbar-item': '/components/issue-564/native-tabbar-item/index',
-    })
-    expect(Object.values(pageJson.usingComponents ?? {})).toContain('/pages/issue-564/index.__scoped-slot-default-0')
-    expect(Object.values(pageJson.usingComponents ?? {})).not.toContain('/pages/issue-564/index.__scoped-slot-default-1')
-    expect(scopedSlotWxml).toContain('<issue-564-native-tabbar-item wx:for="{{__wv_bind_0}}"')
-    expect(scopedSlotWxml).not.toContain('wx:for="{{__wvOwner.tabItems}}"')
-    expect(scopedSlotWxml).toContain('>{{__wv_item_0.label}}</issue-564-native-tabbar-item>')
-    expect(scopedSlotWxml).not.toContain('generic:scoped-slots-default')
-    expect(Object.values(scopedSlotJson.usingComponents ?? {})).toContain('/components/issue-564/native-tabbar-item/index')
-    expect(Object.values(scopedSlotJson.usingComponents ?? {})).not.toContain('/pages/issue-564/index.__scoped-slot-default-1')
-    expect(nativeTabbarJson.componentGenerics?.['scoped-slots-default']?.default).toBe('./__weapp_vite_scoped_slot_generic_component')
-    expect(nativeTabbarWxml).toContain('<slot /><scoped-slots-default')
-    expect(nativeTabbarWxml).toContain('__wvSlotOwnerId="{{__wvSlotOwnerId}}"')
-    expect(nativeTabbarJs).toContain('__wvSlotOwnerId')
-    expect(nativeTabbarGenericJson).toMatchObject({
-      component: true,
-      options: {
-        virtualHost: true,
-      },
-    })
-    expect(await fs.pathExists(nestedScopedSlotWxmlPath)).toBe(false)
-  })
-
-  it('experiment: flex parent keeps projected multi-node slot groups visible via view wrappers', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/slot-flex-layout/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/slot-flex-layout/index.js')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/slot-flex-host/index.wxml')
-
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('slot flex layout experiment')
-    expect(pageWxml).toContain('<view slot="left" class="slot-flex-item slot-flex-item--single-left" data-case="single-left">')
-    expect(pageWxml).toContain('<view slot="middle" class="slot-flex-item slot-flex-item--single-middle" data-case="single-middle">')
-    expect(pageWxml).toContain('<view slot="right" class="slot-flex-item slot-flex-item--single-right" data-case="single-right">')
-    expect(pageWxml).toContain('<weapp-slot-wrapper slot="middle"><view class="slot-flex-item slot-flex-item--middle-multi-center-a" data-case="middle-multi-center-a">')
-    expect(pageWxml).toContain('<weapp-slot-wrapper slot="left"><view class="slot-flex-item slot-flex-item--all-multi-left-a" data-case="all-multi-left-a">')
-    expect(pageWxml).toContain('<weapp-slot-wrapper slot="right"><view class="slot-flex-item slot-flex-item--all-multi-right-a" data-case="all-multi-right-a">')
-    expect(pageWxml).not.toContain('<block slot="middle">')
-    expect(pageWxml).not.toContain('<block slot="left">')
-    expect(pageWxml).not.toContain('<block slot="right">')
-    expect(pageJs).not.toContain('_runE2E')
-    expect(componentWxml).toContain('<slot name="left" />')
-    expect(componentWxml).toContain('<slot name="middle" />')
-    expect(componentWxml).toContain('<slot name="right" />')
-  })
-
-  it('experiment: native slot self-closing and paired tags produce equivalent slot fixtures', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/slot-tag-form/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/slot-tag-form/index.js')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/slot-tag-form/index.json')
-    const selfHostWxmlPath = path.join(DIST_ROOT, 'components/slot-tag-self-host/index.wxml')
-    const pairedHostWxmlPath = path.join(DIST_ROOT, 'components/slot-tag-paired-host/index.wxml')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageJson = await fs.readJson(pageJsonPath)
-    const selfHostWxml = await fs.readFile(selfHostWxmlPath, 'utf-8')
-    const pairedHostWxml = await fs.readFile(pairedHostWxmlPath, 'utf-8')
-
-    expect(selfHostWxml).toContain('<slot name="header" />')
-    expect(selfHostWxml).toContain('<slot />')
-    expect(pairedHostWxml).toContain('<slot name="header"></slot>')
-    expect(pairedHostWxml).toContain('<slot></slot>')
-    expect(pageWxml).toContain('self header: {{sharedHeaderLabel}}')
-    expect(pageWxml).toContain('self body: {{sharedBodyLabel}}')
-    expect(pageWxml).toContain('paired header: {{sharedHeaderLabel}}')
-    expect(pageWxml).toContain('paired body: {{sharedBodyLabel}}')
-    expect(pageJs).toContain('toggleLabels')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJson.usingComponents).toMatchObject({
-      'slot-tag-paired-host': '../../components/slot-tag-paired-host/index',
-      'slot-tag-self-host': '../../components/slot-tag-self-host/index',
-    })
-  })
-
-  it('issue #459: keeps directly imported web-apis polyfills interoperable in github-issues app', async () => {
-    await runBuild()
-
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-459/index.js')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-459/index.wxml')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-459 web-apis polyfill compatibility')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('"/abc", "https://issue-459.invalid"')
-    expect(pageJs).toContain('"bodyValue"')
-    expect(pageJs).toContain('issue-459')
-  })
-
-  it('issue #466: keeps all tdesign Dialog methods callable in github-issues app output', async () => {
-    await runBuild()
-
-    const pageJsPath = path.join(DIST_ROOT, 'subpackages/issue-466/index.js')
-    const nativePageJsPath = path.join(DIST_ROOT, 'subpackages/issue-466/native/index.js')
-    const pageWxmlPath = path.join(DIST_ROOT, 'subpackages/issue-466/index.wxml')
-    const pageJsonPath = path.join(DIST_ROOT, 'subpackages/issue-466/index.json')
-    const dialogIndexPath = path.join(DIST_ROOT, 'subpackages/issue-466/miniprogram_npm/tdesign-miniprogram/dialog/index.js')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const nativePageJs = await fs.readFile(nativePageJsPath, 'utf-8')
-    const pageWxml = stripGeneratedScopeAttrs(await fs.readFile(pageWxmlPath, 'utf-8'))
-    const pageJson = await fs.readJson(pageJsonPath)
-
-    expect(await fs.pathExists(dialogIndexPath)).toBe(true)
-    expect(pageWxml).toContain('issue-466 tdesign Dialog.confirm runtime')
-    expect(pageWxml).toContain('alertType = {{alertType}}')
-    expect(pageWxml).toContain('confirmType = {{confirmType}}')
-    expect(pageWxml).toContain('actionType = {{actionType}}')
-    expect(pageWxml).toContain('closeType = {{closeType}}')
-    expect(pageWxml).toContain('defaultConfirmType = {{defaultConfirmType}}')
-    expect(pageWxml).toContain('<t-dialog id="issue466-dialog" />')
-    expect(pageJson.usingComponents).toMatchObject({
-      't-dialog': 'tdesign-miniprogram/dialog/dialog',
-    })
-    expect(pageJs).toContain('_openAlertE2E')
-    expect(pageJs).toContain('_openDialogE2E')
-    expect(pageJs).toContain('_openConfirmE2E')
-    expect(pageJs).toContain('_openActionE2E')
-    expect(pageJs).toContain('_confirmDialogE2E')
-    expect(pageJs).toContain('_cancelDialogE2E')
-    expect(pageJs).toContain('_selectSecondActionE2E')
-    expect(pageJs).toContain('_prepareCloseHostE2E')
-    expect(pageJs).toContain('_closeDialogE2E')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('_resetE2E')
-    expect(pageJs).toContain('#issue466-dialog')
-    expect(pageJs).toContain('tdesign-miniprogram/dialog/index')
-    expect(pageJs).toContain('issue-466 alert title')
-    expect(pageJs).toContain('issue-466 confirm title')
-    expect(pageJs).toContain('issue-466 action title')
-    expect(pageJs).toContain('issue-466 close title')
-    expect(pageJs).not.toContain('.default.default')
-    expect(pageJs).not.toMatch(/__toESM\([^)]*,\s*1\)/)
-    expect(nativePageJs).toContain('tdesign-miniprogram/dialog/index')
-    expect(nativePageJs).toContain('issue-466 native confirm title')
-    expect(nativePageJs).not.toContain('miniprogram_dist/dialog/index.js')
-    expect(nativePageJs).not.toContain('.default.default')
-    expect(nativePageJs).not.toMatch(/__toESM\([^)]*,\s*1\)/)
-  })
-
-  it('issue #466 computed: keeps build-npm cjs package output stable inside github-issues subpackage', async () => {
-    await runBuild()
-
-    const pageJsPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/index.js')
-    const pageWxmlPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/index.wxml')
-    const pageJsonPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/index.json')
-    const componentJsPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/components/CjsProbe/index.js')
-    const componentWxmlPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/components/CjsProbe/index.wxml')
-    const computedPkgPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/miniprogram_npm/miniprogram-computed/index.js')
-    const computedDialogComponentPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/miniprogram_npm/tdesign-miniprogram/dialog/dialog.js')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJson = await fs.readJson(pageJsonPath)
-    const componentJs = await fs.readFile(componentJsPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-    const computedPkg = await fs.readFile(computedPkgPath, 'utf-8')
-    const computedFastDeepEqualPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/miniprogram_npm/fast-deep-equal/index.js')
-    const computedRfdcPath = path.join(DIST_ROOT, 'subpackages/issue-466-computed/miniprogram_npm/rfdc/index.js')
-
-    expect(await fs.pathExists(computedPkgPath)).toBe(true)
-    expect(await fs.pathExists(computedFastDeepEqualPath)).toBe(true)
-    expect(await fs.pathExists(computedRfdcPath)).toBe(true)
-    expect(await fs.pathExists(computedDialogComponentPath)).toBe(true)
-    expect(pageWxml).toContain('issue-466 computed cjs package')
-    expect(pageWxml).toContain('alertType = {{alertType}}')
-    expect(pageWxml).toContain('confirmType = {{confirmType}}')
-    expect(pageWxml).toContain('actionType = {{actionType}}')
-    expect(pageWxml).toContain('closeType = {{closeType}}')
-    expect(pageWxml).toContain('<issue466-computed-probe')
-    expect(pageWxml).toContain('<t-dialog id="issue466-computed-dialog" />')
-    expect(pageJson.usingComponents).toMatchObject({
-      'issue466-computed-probe': './components/CjsProbe/index',
-      't-dialog': './miniprogram_npm/tdesign-miniprogram/dialog/dialog',
-    })
-    expect(pageJs).toContain('readProbeState')
-    expect(pageJs).toContain('applyNextE2E')
-    expect(pageJs).toContain('_openAlertE2E')
-    expect(pageJs).toContain('_openConfirmE2E')
-    expect(pageJs).toContain('_openActionE2E')
-    expect(pageJs).toContain('_confirmDialogE2E')
-    expect(pageJs).toContain('_cancelDialogE2E')
-    expect(pageJs).toContain('_selectSecondActionE2E')
-    expect(pageJs).toContain('_prepareCloseHostE2E')
-    expect(pageJs).toContain('_closeDialogE2E')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('tdesign-miniprogram/dialog/index')
-    expect(pageJs).toContain('issue-466-computed alert title')
-    expect(pageJs).toContain('issue-466-computed confirm title')
-    expect(pageJs).toContain('issue-466-computed action title')
-    expect(pageJs).toContain('issue-466-computed close title')
-
-    expect(componentWxml).toContain('sum = {{sum}}')
-    expect(componentWxml).toContain('summary = {{summary}}')
-    expect(componentJs).toContain('ComponentWithComputed')
-    expectModuleReference(componentJs, 'miniprogram-computed')
-    expect(componentJs).toContain('watchCount')
-    expect(componentJs).toContain('lastWatch')
-    expect(componentJs).toContain('summary(data)')
-    expect(componentJs).toContain('sum(data)')
-
-    expect(computedPkg).toContain('ComponentWithComputed')
-    expect(computedPkg).toContain('BehaviorWithComputed')
-    expect(computedPkg).toContain('require("rfdc")')
-  })
-
-  it('issue #393: keeps path-mode devDependency chunks out of dist/node_modules', async () => {
-    await runIssue393Build()
-
-    const files = await scanFiles(ISSUE_393_DIST_ROOT)
-    const pageJs = await fs.readFile(path.join(ISSUE_393_DIST_ROOT, 'pages/issue-393/index.js'), 'utf8')
-
-    expect(files).toContain('pages/issue-393/index.js')
-    expect(files).toContain('debounce/index.js')
-    expect(files.some(file => file.startsWith('node_modules/'))).toBe(false)
-    expect(pageJs).toContain('../../debounce/index.js')
-  })
-
-  it('issue #369: does not inject vite client types when weapp.web is not configured', async () => {
-    await runBuild()
-
-    const managedTsconfigPath = path.join(APP_ROOT, '.weapp-vite/tsconfig.app.json')
-    const managedTsconfig = await fs.readJson(managedTsconfigPath)
-
-    expect(managedTsconfig.compilerOptions.types).toContain('weapp-vite/client')
-    expect(managedTsconfig.compilerOptions.types).not.toContain('vite/client')
-  })
-
-  it('issue #380: keeps custom-tab-bar out of default page layouts', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-380/index.wxml')
-    const customTabBarWxmlPath = path.join(DIST_ROOT, 'custom-tab-bar/index.wxml')
-    const customTabBarJsonPath = path.join(DIST_ROOT, 'custom-tab-bar/index.json')
-
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const customTabBarWxml = await fs.readFile(customTabBarWxmlPath, 'utf-8')
-    const customTabBarJson = await fs.readFile(customTabBarJsonPath, 'utf-8')
-
-    expect(pageWxml).toContain(`<weapp-layout-default ${SLOT_OWNER_ATTR}>`)
-    expect(pageWxml).toContain('issue-380 page')
-    expect(customTabBarWxml).toContain('issue-380 custom tab bar')
-    expect(customTabBarWxml).not.toContain('<weapp-layout-default')
-    expect(customTabBarJson).not.toContain('weapp-layout-default')
-  })
-
-  it('issue #385: initializes native default layout state and short-circuits duplicate setPageLayout updates', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-385/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-385/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('<attach-probe id="attach-probe" />')
-    expect(pageWxml).toContain(`!__wv_page_layout_name || __wv_page_layout_name === 'default'`)
-    expect(pageJs).toMatch(/__wv_page_layout_name:\s*[`'"]default[`'"]/)
-    expect(pageJs).toMatch(/__wv_page_layout_props:\s*\{\}/)
-    expect(pageJs).toContain('Object.keys(__wv_current_layout_props).every(')
-    expect(pageJs).toContain('Object.keys(__wv_current_layout_props).length === Object.keys(__wv_next_layout_props).length')
-    expect(pageJs).toContain('this.setData({')
-  })
-
-  it('issue #389: keeps native setPageLayout off the wevu layout runtime path', async () => {
-    await runBuild()
-
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-389/index.js')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageJs).toContain('issue-389 native runtime export should stay native-only')
-    expect(pageJs).toContain('__wevuSetPageLayout')
-    expect(pageJs).not.toContain('__wevuPageLayoutState')
-    expect(pageJs).not.toContain('usePageLayout() 必须在 setup() 的同步阶段调用')
-    expect(pageJs).not.toContain('syncRuntimePageLayoutStateFromRuntime')
-  })
-
-  it('issue #398: emits layout and component importers that share the same wevu runtime chunk', async () => {
-    await runBuild()
-
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-398/index.js')
-    const layoutJsPath = path.join(DIST_ROOT, 'layouts/issue-398-shell.js')
-    const navbarJsPath = path.join(DIST_ROOT, 'components/issue-398/BaseNavbar/index.js')
-    const footerJsPath = path.join(DIST_ROOT, 'components/issue-398/BaseFooter/index.js')
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-398/index.wxml')
-
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const layoutJs = await fs.readFile(layoutJsPath, 'utf-8')
-    const navbarJs = await fs.readFile(navbarJsPath, 'utf-8')
-    const footerJs = await fs.readFile(footerJsPath, 'utf-8')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain(`<weapp-layout-issue-398-shell ${SLOT_OWNER_ATTR}>`)
-    expect(pageWxml).toContain('{{issue398Title}}')
-    expect(pageWxml).toContain('{{issue398PageMarker}}')
-    expect(pageWxml).toContain('{{issue398TapLabel}}')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('issue398NavbarLabel')
-    expect(pageJs).toContain('issue398FooterLabel')
-    expect(layoutJs).not.toContain('../components/issue-398/BaseNavbar/index.vue')
-    expect(layoutJs).not.toContain('../components/issue-398/BaseFooter/index.vue')
-    expect(navbarJs).toContain('issue-398 navbar')
-    expect(navbarJs).toContain('__issue398NavbarMounted')
-    expect(navbarJs).toContain('__issue398NavbarLabel')
-    expect(footerJs).toContain('issue-398 footer')
-    expect(footerJs).toContain('__issue398FooterMounted')
-    expect(footerJs).toContain('__issue398FooterLabel')
-
-    const pageSharedRuntimePath = resolveSharedRuntimeImport(pageJsPath, pageJs)
-    const layoutSharedRuntimePath = resolveSharedRuntimeImport(layoutJsPath, layoutJs)
-    const navbarSharedRuntimePath = resolveSharedRuntimeImport(navbarJsPath, navbarJs)
-    const footerSharedRuntimePath = resolveSharedRuntimeImport(footerJsPath, footerJs)
-
-    expect(pageSharedRuntimePath).toBeTruthy()
-    expect(layoutSharedRuntimePath).toBeTruthy()
-    expect(navbarSharedRuntimePath).toBeTruthy()
-    expect(footerSharedRuntimePath).toBeTruthy()
-    expect(pageSharedRuntimePath).toBe(layoutSharedRuntimePath)
-    expect(pageSharedRuntimePath).toBe(navbarSharedRuntimePath)
-    expect(pageSharedRuntimePath).toBe(footerSharedRuntimePath)
-    expect(await fs.pathExists(pageSharedRuntimePath!)).toBe(true)
-  })
-
-  it('issue #404: keeps onPageScroll page hooks enabled and exposes runtime probes in the page bundle', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-404/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-404/index.js')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-404 onPageScroll bridge')
-    expect(pageWxml).toContain('has instance onPageScroll')
-    expect(pageWxml).toContain('latest scrollTop')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('hasInstanceOnPageScroll')
-    expect(pageJs).toContain('enableOnPageScroll')
-  })
-
-  it('issue #418/#419: keeps native template refs stable for third-party components', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-418-419/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-418-419/index.js')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-418-419/index.json')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageJson = await fs.readFile(pageJsonPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-418-419 template ref native component')
-    expect(pageWxml).toContain('<native-ref-probe')
-    expect(pageWxml).toContain('issue-418-419')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('nativeButtonRef')
-    expect(pageJs).toContain('descriptorConfigurable')
-    expect(pageJson).toContain('../../components/issue-418-419/NativeRefProbe/index')
-  })
-
-  it('issue #446: supports template refs together with Vue 3.4 shortBind syntax', async () => {
-    await runBuild()
-
-    const pageWxmlPath = path.join(DIST_ROOT, 'pages/issue-446/index.wxml')
-    const pageJsPath = path.join(DIST_ROOT, 'pages/issue-446/index.js')
-    const pageJsonPath = path.join(DIST_ROOT, 'pages/issue-446/index.json')
-    const componentWxmlPath = path.join(DIST_ROOT, 'components/issue-446/ShortBindProbe/index.wxml')
-    const pageWxml = await fs.readFile(pageWxmlPath, 'utf-8')
-    const pageJs = await fs.readFile(pageJsPath, 'utf-8')
-    const pageJson = await fs.readFile(pageJsonPath, 'utf-8')
-    const componentWxml = await fs.readFile(componentWxmlPath, 'utf-8')
-
-    expect(pageWxml).toContain('issue-446 template ref and shortBind')
-    expect(pageWxml).toContain('__wv-ref-0')
-    expect(pageWxml).toContain('visible="{{visible}}"')
-    expect(pageWxml).toContain('foo-bar="{{fooBar}}"')
-    expect(pageJs).toContain('_runE2E')
-    expect(pageJs).toContain('issue-446-short-bind')
-    expect(pageJs).toContain('__wevuTemplateRefs')
-    expect(pageJs).toContain('name: "nativeAnchor"')
-    expect(pageJs).toContain('name: "shortBindProbe"')
-    expect(pageJson).toContain('"short-bind-probe": "/components/issue-446/ShortBindProbe/index"')
-    expect(componentWxml).toContain('{{props.visible ? \'visible\' : \'hidden\'}}')
-    expect(componentWxml).toContain('{{props.fooBar}}')
-  })
-
-  it('issue #289: compiles split pages with per-page controls and safe class bindings', async () => {
-    await runBuild()
-
-    const navPageWxmlPath = path.join(DIST_ROOT, 'pages/issue-289/index.wxml')
-    const navPageJsPath = path.join(DIST_ROOT, 'pages/issue-289/index.js')
-    const objectPageWxmlPath = path.join(DIST_ROOT, 'pages/issue-289/object-literal/index.wxml')
-    const objectPageJsPath = path.join(DIST_ROOT, 'pages/issue-289/object-literal/index.js')
-    const mapPageWxmlPath = path.join(DIST_ROOT, 'pages/issue-289/map-class/index.wxml')
-    const mapPageJsPath = path.join(DIST_ROOT, 'pages/issue-289/map-class/index.js')
-    const rootPageWxmlPath = path.join(DIST_ROOT, 'pages/issue-289/root-class/index.wxml')
-    const rootPageJsPath = path.join(DIST_ROOT, 'pages/issue-289/root-class/index.js')
-    const computedPageWxmlPath = path.join(DIST_ROOT, 'pages/issue-289/computed-class/index.wxml')
-    const computedPageJsPath = path.join(DIST_ROOT, 'pages/issue-289/computed-class/index.js')
-
-    const objectLiteralWxmlPath = path.join(DIST_ROOT, 'components/issue-289/ObjectLiteralExample/index.wxml')
-    const objectLiteralJsPath = path.join(DIST_ROOT, 'components/issue-289/ObjectLiteralExample/index.js')
-    const mapClassWxmlPath = path.join(DIST_ROOT, 'components/issue-289/MapClassExample/index.wxml')
-    const mapClassJsPath = path.join(DIST_ROOT, 'components/issue-289/MapClassExample/index.js')
-    const rootClassWxmlPath = path.join(DIST_ROOT, 'components/issue-289/RootClassExample/index.wxml')
-    const rootClassJsPath = path.join(DIST_ROOT, 'components/issue-289/RootClassExample/index.js')
-    const computedClassWxmlPath = path.join(DIST_ROOT, 'components/issue-289/ComputedClassExample/index.wxml')
-    const computedClassJsPath = path.join(DIST_ROOT, 'components/issue-289/ComputedClassExample/index.js')
-
-    const navPageWxml = await fs.readFile(navPageWxmlPath, 'utf-8')
-    const navPageJs = await fs.readFile(navPageJsPath, 'utf-8')
-    const objectPageWxml = await fs.readFile(objectPageWxmlPath, 'utf-8')
-    const objectPageJs = await fs.readFile(objectPageJsPath, 'utf-8')
-    const mapPageWxml = await fs.readFile(mapPageWxmlPath, 'utf-8')
-    const mapPageJs = await fs.readFile(mapPageJsPath, 'utf-8')
-    const rootPageWxml = await fs.readFile(rootPageWxmlPath, 'utf-8')
-    const rootPageJs = await fs.readFile(rootPageJsPath, 'utf-8')
-    const computedPageWxml = await fs.readFile(computedPageWxmlPath, 'utf-8')
-    const computedPageJs = await fs.readFile(computedPageJsPath, 'utf-8')
-
-    const objectLiteralWxml = await fs.readFile(objectLiteralWxmlPath, 'utf-8')
-    const objectLiteralJs = await fs.readFile(objectLiteralJsPath, 'utf-8')
-    const mapClassWxml = await fs.readFile(mapClassWxmlPath, 'utf-8')
-    const mapClassJs = await fs.readFile(mapClassJsPath, 'utf-8')
-    const rootClassWxml = await fs.readFile(rootClassWxmlPath, 'utf-8')
-    const rootClassJs = await fs.readFile(rootClassJsPath, 'utf-8')
-    const computedClassWxml = await fs.readFile(computedClassWxmlPath, 'utf-8')
-    const computedClassJs = await fs.readFile(computedClassJsPath, 'utf-8')
-
-    expect(navPageWxml).toContain('wx:for="{{sceneLinks}}"')
-    expect(navPageWxml).toContain('wx:key="url"')
-    expect(navPageWxml).toContain('url="{{item.url}}"')
-    expect(navPageJs).toContain('sceneLinks')
-    expect(navPageJs).toContain('/pages/issue-289/object-literal/index')
-    expect(navPageJs).toContain('/pages/issue-289/map-class/index')
-    expect(navPageJs).toContain('/pages/issue-289/root-class/index')
-    expect(navPageJs).toContain('/pages/issue-289/computed-class/index')
-
-    expect(objectPageWxml).toContain('<object-literal-example')
-    expect(objectPageWxml).toContain('show-list="{{controlState.showList}}"')
-    expect(objectPageWxml).toContain('compact-mode="{{controlState.compactMode}}"')
-    expect(objectPageWxml).toContain('active-id="{{activeId || \'\'}}"')
-    expect(objectPageWxml).toContain('issue289-object-toggle-list')
-    expect(objectPageWxml).toContain('issue289-object-toggle-compact')
-    expect(objectPageWxml).toContain('issue289-object-cycle-active')
-    expect(objectPageWxml).toContain('bindtap="__weapp_vite_inline"')
-    expect(objectPageWxml).toContain('data-wi-tap=')
-    expect(objectPageJs).toContain('controlState')
-    expect(objectPageJs).toContain('.showList')
-    expect(objectPageJs).toContain('.compactMode')
-    expect(objectPageJs).toContain('toggleShowList')
-    expect(objectPageJs).toContain('toggleCompactMode')
-    expect(objectPageJs).toContain('cycleActive')
-    expect(objectPageJs).toContain('runE2E')
-    expect(objectPageJs).toContain('showListRoundTripWorked')
-
-    expect(mapPageWxml).toContain('<map-class-example')
-    expect(mapPageWxml).toContain('callout-expanded="{{controlState.calloutExpanded}}"')
-    expect(mapPageWxml).toContain('show-callout-list="{{controlState.showCalloutList}}"')
-    expect(mapPageWxml).toContain('selected-event-idx="{{controlState.selectedIndex}}"')
-    expect(mapPageWxml).toContain('气泡尺寸')
-    expect(mapPageWxml).toContain('气泡列表')
-    expect(mapPageWxml).toContain('选中事件')
-    expect(mapPageWxml).toContain('bindtap="__weapp_vite_inline"')
-    expect(mapPageWxml).toContain('data-wi-tap=')
-    expect(mapPageJs).toContain('controlState')
-    expect(mapPageJs).toContain('.calloutExpanded')
-    expect(mapPageJs).toContain('.showCalloutList')
-    expect(mapPageJs).toContain('.selectedIndex')
-    expect(mapPageJs).toContain('toggleCalloutExpanded')
-    expect(mapPageJs).toContain('toggleShowCalloutList')
-    expect(mapPageJs).toContain('cycleSelectedEvent')
-    expect(mapPageJs).toContain('runE2E')
-    expect(mapPageJs).toContain('showCalloutListChanged')
-
-    expect(rootPageWxml).toContain('<root-class-example')
-    expect(rootPageWxml).toContain('show-options="{{controlState.showOptions}}"')
-    expect(rootPageWxml).toContain('selected-option-id="{{selectedOptionId || \'\'}}"')
-    expect(rootPageWxml).toContain('issue289-root-toggle-options')
-    expect(rootPageWxml).toContain('issue289-root-cycle-option')
-    expect(rootPageWxml).toContain('bindtap="__weapp_vite_inline"')
-    expect(rootPageWxml).toContain('data-wi-tap=')
-    expect(rootPageJs).toContain('controlState')
-    expect(rootPageJs).toContain('.showOptions')
-    expect(rootPageJs).toContain('toggleShowOptions')
-    expect(rootPageJs).toContain('cycleOption')
-    expect(rootPageJs).toContain('runE2E')
-    expect(rootPageJs).toContain('selectedIndexChanged')
-
-    expect(computedPageWxml).toContain('<computed-class-example')
-    expect(computedPageWxml).toContain('source-enabled="{{controlState.sourceEnabled}}"')
-    expect(computedPageWxml).toContain('show-items="{{controlState.showItems}}"')
-    expect(computedPageWxml).toContain('selected-index="{{controlState.selectedIndex}}"')
-    expect(computedPageWxml).toContain('source：')
-    expect(computedPageWxml).toContain('选中项')
-    expect(computedPageWxml).toContain('bindtap="__weapp_vite_inline"')
-    expect(computedPageWxml).toContain('data-wi-tap=')
-    expect(computedPageJs).toContain('controlState')
-    expect(computedPageJs).toContain('.sourceEnabled')
-    expect(computedPageJs).toContain('.showItems')
-    expect(computedPageJs).toContain('.selectedIndex')
-    expect(computedPageJs).toContain('toggleSourceEnabled')
-    expect(computedPageJs).toContain('toggleShowItems')
-    expect(computedPageJs).toContain('cycleSelected')
-    expect(computedPageJs).toContain('runE2E')
-    expect(computedPageJs).toContain('showItemsChanged')
-
-    expect(objectLiteralWxml).toContain('root="{{ {a:\'aaaa\'} }}"')
-    expect(objectLiteralWxml).not.toContain('root="{{{')
-    expect(objectLiteralWxml).not.toContain('root="{{({')
-    expect(objectLiteralWxml).not.toMatch(/root="\{\{__wv_bind_\d+\}\}"/)
-    expect(objectLiteralWxml).toMatch(/wx:if="\{\{showList\}\}"/)
-    expect(objectLiteralWxml).toMatch(/wx:for="\{\{items\}\}"/)
-    expect(objectLiteralWxml).toContain('wx:else')
-
-    const objectClassBindingTokens = objectLiteralWxml.match(/__wv_cls_\d+/g) ?? []
-    expect(new Set(objectClassBindingTokens).size).toBeGreaterThanOrEqual(2)
-
-    expect(mapClassWxml).toMatch(/wx:for-index="(?:__wv_index_0|index)"/)
-    expect(mapClassWxml).toMatch(/__wv_cls_\d+\[(?:__wv_index_0|index)\]/)
-    expect(mapClassWxml).toContain('min-scale="{{3}}"')
-    expect(mapClassWxml).toContain('max-scale="{{20}}"')
-    expect(mapClassWxml).toContain('bindmarkertap="__weapp_vite_inline"')
-    expect(mapClassWxml).toContain('data-wi-markertap="i')
-    expect(mapClassWxml).toContain('bindregionchange="__weapp_vite_inline"')
-    expect(mapClassWxml).toContain('data-wi-regionchange="i')
-    expect(mapClassWxml).toContain('show-compass="{{true}}"')
-    expect(mapClassWxml).toContain('enable-zoom="{{true}}"')
-    expect(mapClassWxml).toMatch(/wx:for="\{\{events\}\}"/)
-    expect(mapClassWxml).toMatch(/wx:for="\{\{mapMetaList\}\}"/)
-    expect(mapClassWxml).toMatch(/wx:if="\{\{showCalloutList\}\}"/)
-
-    const mapClassBindingTokens = mapClassWxml.match(/__wv_cls_\d+/g) ?? []
-    expect(new Set(mapClassBindingTokens).size).toBeGreaterThanOrEqual(1)
-
-    expect(objectLiteralJs).not.toMatch(/__wv_bind_\d+\s*=\s*\(\s*\(\)\s*=>\s*\(\{\s*a:\s*['"`]aaaa['"`]\s*\}\)\s*\)/)
-    expect(objectLiteralJs).not.toMatch(/return\s*\{\s*a:\s*['"`]aaaa['"`]\s*\}/)
-    expect(objectLiteralJs).toContain('showList')
-    expect(objectLiteralJs).toContain('compactMode')
-    expect(objectLiteralJs).toContain('activeId')
-
-    expect(mapClassJs).toContain('selectedEventIdx')
-    expect(mapClassJs).toContain('isPublic')
-    expect(mapClassJs).toContain('includePoints')
-    expect(mapClassJs).toContain('polyline')
-    expect(mapClassJs).toContain('circles')
-    expect(mapClassJs).toContain('calloutExpanded')
-    expect(mapClassJs).toContain('showCalloutList')
-    expect(mapClassJs).toContain('safeSelectedEventIdx')
-    expect(mapClassJs).toContain('mapMetaList')
-
-    expect(rootClassWxml).toMatch(/class="\{\{__wv_cls_\d+\}\}"/)
-    expect(rootClassWxml).toMatch(/wx:if="\{\{showOptions\}\}"/)
-    expect(rootClassWxml).toMatch(/wx:for="\{\{options\}\}"/)
-    expect(rootClassJs).toContain('selectedClassName')
-    expect(rootClassJs).toContain('showOptions')
-    expect(rootClassJs).toContain('selectedOptionId')
-
-    expect(computedClassWxml).toMatch(/class="\{\{__wv_cls_\d+\}\}"/)
-    expect(computedClassWxml).toMatch(/wx:if="\{\{showItems\}\}"/)
-    expect(computedClassWxml).toMatch(/wx:for="\{\{items\}\}"/)
-    expect(computedClassJs).toContain('computedValue')
-    expect(computedClassJs).toContain('sourceEnabled')
-    expect(computedClassJs).toContain('showItems')
-    expect(computedClassJs).toContain('selectedIndex')
-    expect(computedClassJs).toContain('safeSelectedIndex')
-    expect(computedClassJs).toContain('computedListClass')
-  })
-  it('issue #294: injects share lifecycle options for script-setup page hooks', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-294/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-294/index.js')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-    const issuePageJsonPath = path.join(DIST_ROOT, 'pages/issue-294/index.json')
-    const issuePageJson = await fs.readFile(issuePageJsonPath, 'utf-8')
-    const { code: wevuRuntime } = await findWevuVendorChunkContaining(DIST_ROOT, ['showShareMenu'])
-
-    expect(issuePageWxml).toContain('issue-294 share hooks')
-    expect(issuePageJs).toContain('enableOnShareAppMessage: true')
-    expect(issuePageJs).toContain('enableOnShareTimeline: true')
-    expect(issuePageJs).not.toMatch(/onShareAppMessage(?:\(\)|:function\(\))\{return\{\}\}/)
-    expect(issuePageJs).not.toMatch(/onShareTimeline(?:\(\)|:function\(\))\{return\{\}\}/)
-    expect(issuePageJs).toContain('issue-294-share-')
-    expect(issuePageJs).toContain('issue-294-timeline-')
-    expect(issuePageJson).not.toContain('"enableShareAppMessage"')
-    expect(issuePageJson).not.toContain('"enableShareTimeline"')
-    expect(wevuRuntime).toContain('showShareMenu')
-  })
-  it('issue #297: compiles complex call expressions', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-297/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-297/index.js')
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-297 complex call expressions')
-    expect(issuePageWxml).toContain('Case A · v-bind 调用表达式')
-    expect(issuePageWxml).toContain('Case B · v-if + v-for 调用表达式')
-    expect(issuePageWxml).toContain('Case C · 多参数调用（含激活项）')
-    expect(issuePageWxml).toContain('Case A(v-bind):')
-    expect(issuePageWxml).toContain('Case B(v-if/v-for):')
-    expect(issuePageWxml).toContain('Case C(多参数):')
-    expect(issuePageWxml).toContain('新增列表项')
-    expect(issuePageWxml).toContain('重置列表')
-    expect(issuePageWxml).toMatch(/wx:if="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/wx:for="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-title="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-hello="\{\{__wv_bind_\d+\[[^\]]+\]\}\}"/)
-    expect(issuePageWxml).toMatch(/\{\{__wv_bind_\d+\[[^\]]+\]\}\}/)
-    expect(issuePageWxml).not.toContain('sayHello(')
-    const bindTokens = issuePageWxml.match(/__wv_bind_\d+/g) ?? []
-    expect(new Set(bindTokens).size).toBeGreaterThanOrEqual(5)
-
-    expect(issuePageJs).toContain('sayHello(1, \'root\', dasd)')
-    expect(issuePageJs).toContain('dasd')
-    expect(issuePageJs).toContain('this.sayHello')
-    expect(issuePageJs).toContain('_runE2E')
-  })
-
-  it('issue #297: setup method call variants compile to stable bindings', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-297-setup-method-calls/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-297-setup-method-calls/index.js')
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-297 setup method call variants')
-    expect(issuePageWxml).toContain('Case A · 插值调用 + 同级静态文本 + 同级元素')
-    expect(issuePageWxml).toContain('Case B · v-bind 多参数调用')
-    expect(issuePageWxml).toContain('Case C · v-if + v-for 调用表达式')
-    expect(issuePageWxml).toContain('Case D · 成员调用 / 模板字符串 / 三元表达式')
-    expect(issuePageWxml).toContain('Case E · 可选调用 + 空值兜底')
-    expect(issuePageWxml).toMatch(/wx:if="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/wx:for="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-inline="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-multi="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-label="\{\{__wv_bind_\d+\[[^\]]+\]\}\}"/)
-    expect(issuePageWxml).toMatch(/data-loop="\{\{__wv_bind_\d+\[[^\]]+\]\}\}"/)
-    expect(issuePageWxml).toMatch(/data-member="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-template="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-ternary="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-wrap="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/data-optional="\{\{__wv_bind_\d+\}\}"/)
-    expect(issuePageWxml).not.toContain('getCase()')
-    expect(issuePageWxml).not.toContain('sayCase(')
-    expect(issuePageWxml).not.toContain('getOptionalInvoker?.(')
-
-    expect(issuePageJs).toContain('this.getCase')
-    expect(issuePageJs).toContain('this.sayCase')
-    expect(issuePageJs).toContain('this.getRows')
-    expect(issuePageJs).toContain('this.getOptionalInvoker')
-    expect(issuePageJs).toContain('_runE2E')
-  })
-
-  it('issue #302: keeps v-for class bindings in sync with active state', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-302/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-302/index.js')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-302 v-for class binding update')
-    expect(issuePageWxml).toContain('active: {{active}}')
-    expect(issuePageWxml).toContain('wx:for="{{tabs}}"')
-    expect(issuePageWxml).toContain('wx:key="id"')
-    expect(issuePageWxml).toMatch(/wx:for-index="(?:__wv_index_0|index)"/)
-    expect(issuePageWxml).toMatch(/class="\{\{__wv_cls_\d+\[(?:__wv_index_0|index)\]\}\}"/)
-    expect(issuePageWxml).not.toContain('active === tab.id')
-
-    const classBindingTokens = issuePageWxml.match(/__wv_cls_\d+/g) ?? []
-    expect(new Set(classBindingTokens).size).toBeGreaterThanOrEqual(1)
-
-    expect(issuePageJs).toContain('setActive')
-    expect(issuePageJs).toContain('active')
-    expect(issuePageJs).toContain('tabs')
-    expect(issuePageJs).toContain('_runE2E')
-    expect(issuePageJs).toContain('issue302-item-')
-    expect(issuePageJs).toContain('issue302-item-active')
-    expect(issuePageJs).toContain('issue302-item-inactive')
-  })
-
-  it('issue #309: keeps onLoad hook active even without onPullDownRefresh', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-309/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-309/index.js')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-309 onLoad hook')
-    expect(issuePageWxml).toContain('loadCount: {{loadCount}}')
-    expect(issuePageJs).toContain('_runE2E')
-    expect(issuePageJs).toContain('__wevu_isPage: true')
-    expect(issuePageJs).toContain('loadCount')
-    expect(issuePageJs).not.toContain('onPullDownRefresh')
-  })
-
-  it('issue #309: keeps onLoad hook active with created setupLifecycle', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-309-created/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-309-created/index.js')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-309 created lifecycle onLoad hook')
-    expect(issuePageWxml).toContain('loadCount: {{loadCount}}')
-    expect(issuePageJs).toContain('_runE2E')
-    expect(issuePageJs).toContain('__wevu_isPage: true')
-    expect(issuePageJs).toContain('loadCount')
-    expect(issuePageJs).not.toContain('onPullDownRefresh')
-  })
-
-  it('issue #312: keeps setup computed object bindings reactive after reference round trip', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-312/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-312/index.js')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-312 computed object round trip')
-    expect(issuePageWxml).toContain('current option:')
-    expect(issuePageWxml).toContain('issue312-probe')
-    expect(issuePageWxml).toContain('data-current-label="{{option.label}}"')
-    expect(issuePageWxml).toContain('issue312-btn-inc')
-    expect(issuePageWxml).toContain('issue312-btn-dec')
-    expect(issuePageJs).toContain('_runE2E')
-    expect(issuePageJs).toContain('options')
-    expect(issuePageJs).toContain('option')
-    expect(issuePageJs).toContain('index')
-  })
-
-  it('issue #316: compiles kebab-case component events to colon-prefixed bindings', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-316/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-316/index.js')
-    const emitterWxmlPath = path.join(DIST_ROOT, 'components/issue-316/HyphenEventEmitter/index.wxml')
-    const emitterJsPath = path.join(DIST_ROOT, 'components/issue-316/HyphenEventEmitter/index.js')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-    const emitterWxml = await fs.readFile(emitterWxmlPath, 'utf-8')
-    const emitterJs = await fs.readFile(emitterJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-316 hyphen event binding')
-    expect(issuePageWxml).toContain('custom component event: overlay-click')
-    expect(issuePageWxml).toContain('bind:overlay-click="__weapp_vite_inline"')
-    expect(issuePageWxml).toContain('data-wd-overlay-click="1"')
-    expect(issuePageWxml).toContain('data-wi-overlay-click="i0"')
-    expect(issuePageWxml).toContain('data-overlay-count="{{overlayClickCount}}"')
-    expect(issuePageWxml).not.toContain('bindoverlay-click=')
-    expect(issuePageJs).toContain('handleOverlayClick')
-    expect(issuePageJs).toContain('_runE2E')
-
-    expect(emitterWxml).toContain('issue316-emitter-btn')
-    expect(emitterWxml).toContain('bindtap="__weapp_vite_inline"')
-    expect(emitterWxml).toContain('data-wi-tap=')
-    expect(emitterJs).toContain('emitOverlayClick')
-  })
-
-  it('issue #318: auto injects setData.pick for template-used keys', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-318/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-318/index.js')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-318 auto setData pick from template')
-    expect(issuePageWxml).toContain('wx:for="{{list}}"')
-    expect(issuePageWxml).toContain('count: {{count}}')
-    expect(issuePageWxml).toContain('size: {{list.length}}')
-    expect(issuePageWxml).toContain('data-line="{{__wv_bind_')
-    expect(issuePageWxml).toContain('data-meta="{{__wv_bind_')
-    expect(issuePageWxml).toContain('{{__wv_bind_')
-    expect(issuePageWxml).not.toContain('formatRow(')
-    expect(issuePageWxml).not.toContain('formatMeta(')
-
-    const bindTokens = issuePageWxml.match(/__wv_bind_\d+/g) ?? []
-    expect(new Set(bindTokens).size).toBeGreaterThanOrEqual(2)
-
-    expect(issuePageJs).toContain('_runE2E')
-    expect(issuePageJs).toContain('setData')
-    expect(issuePageJs).toContain('pick')
-    expect(issuePageJs).toContain('"count"')
-    expect(issuePageJs).toContain('"list"')
-    expect(issuePageJs).toMatch(/['"`]__wv_bind_\d+['"`]/)
-  })
-
-  it('issue #322: keeps static class and hidden v-show fallback when expression access throws', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-322/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-322/index.js')
-    const issuePageWxml = stripGeneratedScopeAttrs(await fs.readFile(issuePageWxmlPath, 'utf-8'))
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-322 class/v-show first paint flicker')
-    expect(issuePageWxml).toContain('set email error')
-    expect(issuePageWxml).toContain('clear email error')
-    expect(issuePageWxml).toMatch(/class="\{\{__wv_cls_\d+\}\}"/)
-    expect(issuePageWxml).toMatch(/style="\{\{__wv_style_\d+\}\}"/)
-    expect(issuePageWxml).not.toContain('errors.email ?')
-    expect(issuePageWxml).toMatch(/<input[^>]*class="\{\{__wv_cls_\d+\}\}"[^>]*\/>/)
-    expect(issuePageWxml).not.toContain('</input>')
-    expect(issuePageWxml).toMatch(/<input[^>]*\/><view style="\{\{__wv_style_\d+\}\}" class="issue322-error-tip">/)
-
-    expect(issuePageJs).toMatch(/__wv_cls_\d+/)
-    expect(issuePageJs).toMatch(/__wv_style_\d+/)
-    expect(issuePageJs).toContain('hasEmailError: Boolean(errors.value?.email)')
-    expect(issuePageJs).toContain('emailError: errors.value?.email ?? ""')
-    expect(issuePageJs).toMatch(/return\s+["'`]issue322-input issue322-input-base["'`]/)
-    expect(issuePageJs).toMatch(/return\s+["'`]display: none["'`]/)
-  })
-
-  it('issue #317: keeps shared chunk duplication in subpackages without invalid self or runtime imports', async () => {
-    await runBuild()
-
-    const itemSharedPath = path.join(DIST_ROOT, 'subpackages/item/weapp-shared/common.js')
-    const userSharedPath = path.join(DIST_ROOT, 'subpackages/user/weapp-shared/common.js')
-    const itemRuntimePath = path.join(DIST_ROOT, 'subpackages/item/rolldown-runtime.js')
-    const userRuntimePath = path.join(DIST_ROOT, 'subpackages/user/rolldown-runtime.js')
-    const fallbackUnderscorePath = path.join(DIST_ROOT, 'subpackages_item_subpackages_user/common.js')
-    const fallbackPlusPath = path.join(DIST_ROOT, 'subpackages_item+subpackages_user/common.js')
-
-    const itemShared = await fs.readFile(itemSharedPath, 'utf-8')
-    const userShared = await fs.readFile(userSharedPath, 'utf-8')
-    const itemVendorRuntimePath = resolveSharedRuntimeImport(itemSharedPath, itemShared)
-    const userVendorRuntimePath = resolveSharedRuntimeImport(userSharedPath, userShared)
-
-    expect(itemShared).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/rolldown-runtime\.js\1\)|from\s+(['"`])\.\.\/\.\.\/rolldown-runtime\.js\2)/)
-    expect(userShared).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/rolldown-runtime\.js\1\)|from\s+(['"`])\.\.\/\.\.\/rolldown-runtime\.js\2)/)
-    expect(itemShared).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/common\.js\1\)|from\s+(['"`])\.\.\/\.\.\/common\.js\2)/)
-    expect(userShared).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/common\.js\1\)|from\s+(['"`])\.\.\/\.\.\/common\.js\2)/)
-
-    expect(itemShared).not.toMatch(/require\((['"`]).*subpackages_item.*subpackages_user\/common\.js\1\)/)
-    expect(userShared).not.toMatch(/require\((['"`]).*subpackages_item.*subpackages_user\/common\.js\1\)/)
-    expect(itemVendorRuntimePath).toBeDefined()
-    expect(userVendorRuntimePath).toBeDefined()
-    expect(itemVendorRuntimePath).toContain(`${path.sep}weapp-vendors${path.sep}`)
-    expect(userVendorRuntimePath).toContain(`${path.sep}weapp-vendors${path.sep}`)
-
-    expect(await fs.pathExists(fallbackUnderscorePath)).toBe(false)
-    expect(await fs.pathExists(fallbackPlusPath)).toBe(false)
-    expect(await fs.pathExists(itemRuntimePath)).toBe(true)
-    expect(await fs.pathExists(userRuntimePath)).toBe(true)
-  })
-
-  it('issue #340: keeps cross-subpackage source imports runnable for item/login-required and user/register/form', async () => {
-    await runBuild()
-
-    const appJsonPath = path.join(DIST_ROOT, 'app.json')
-    const itemPageJsPath = path.join(DIST_ROOT, 'subpackages/item/login-required/index.js')
-    const userPageJsPath = path.join(DIST_ROOT, 'subpackages/user/register/form.js')
-    const invalidSharedPageWxmlPath = path.join(DIST_ROOT, 'subpackages/item/issue-340-shared.wxml')
-    const itemSharedPath = path.join(DIST_ROOT, 'subpackages/item/weapp-shared/common.js')
-    const userSharedPath = path.join(DIST_ROOT, 'subpackages/user/weapp-shared/common.js')
-    const itemInvalidCommonPath = path.join(DIST_ROOT, 'subpackages/item/common.js')
-    const userInvalidCommonPath = path.join(DIST_ROOT, 'subpackages/user/common.js')
-    const itemVendorsPath = path.join(DIST_ROOT, 'subpackages/item/vendors.js')
-    const userVendorsPath = path.join(DIST_ROOT, 'subpackages/user/vendors.js')
-    const itemRuntimePath = path.join(DIST_ROOT, 'subpackages/item/rolldown-runtime.js')
-    const userRuntimePath = path.join(DIST_ROOT, 'subpackages/user/rolldown-runtime.js')
-
-    const appJson = JSON.parse(await fs.readFile(appJsonPath, 'utf-8'))
-    const itemPageJs = await fs.readFile(itemPageJsPath, 'utf-8')
-    const userPageJs = await fs.readFile(userPageJsPath, 'utf-8')
-    const itemShared = await fs.readFile(itemSharedPath, 'utf-8')
-    const userShared = await fs.readFile(userSharedPath, 'utf-8')
-    const itemVendorRuntimePath = resolveSharedRuntimeImport(itemSharedPath, itemShared)
-    const userVendorRuntimePath = resolveSharedRuntimeImport(userSharedPath, userShared)
-    const itemSubPackage = (appJson.subPackages ?? appJson.subpackages ?? []).find((entry: any) => entry?.root === 'subpackages/item')
-
-    expect(itemPageJs).toContain('item-login-required:issue-340:shared')
-    expect(userPageJs).toContain('user-register-form:issue-340:shared')
-    expect(itemPageJs).toMatch(/(?:require\((['"`])\.\.\/weapp-shared\/common(?:\.\d+)?\.js\1\)|from\s+(['"`])\.\.\/weapp-shared\/common(?:\.\d+)?\.js\2)/)
-    expect(userPageJs).toMatch(/(?:require\((['"`])\.\.\/weapp-shared\/common(?:\.\d+)?\.js\1\)|from\s+(['"`])\.\.\/weapp-shared\/common(?:\.\d+)?\.js\2)/)
-    expect(itemPageJs).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/\.\.\/common\.js\1\)|from\s+(['"`])\.\.\/\.\.\/\.\.\/common\.js\2)/)
-    expect(userPageJs).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/\.\.\/common\.js\1\)|from\s+(['"`])\.\.\/\.\.\/\.\.\/common\.js\2)/)
-    expect(itemPageJs).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/common(?:\.\d+)?\.js\1\)|from\s+(['"`])\.\.\/\.\.\/common(?:\.\d+)?\.js\2)/)
-    expect(userPageJs).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/common(?:\.\d+)?\.js\1\)|from\s+(['"`])\.\.\/\.\.\/common(?:\.\d+)?\.js\2)/)
-    expect(itemPageJs).not.toMatch(/vendors(?:\.\d+)?\.js/)
-    expect(userPageJs).not.toMatch(/vendors(?:\.\d+)?\.js/)
-
-    expect(itemShared).toContain('issue-340')
-    expect(userShared).toContain('issue-340')
-    expect(itemShared).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/rolldown-runtime\.js\1\)|from\s+(['"`])\.\.\/\.\.\/rolldown-runtime\.js\2)/)
-    expect(userShared).not.toMatch(/(?:require\((['"`])\.\.\/\.\.\/rolldown-runtime\.js\1\)|from\s+(['"`])\.\.\/\.\.\/rolldown-runtime\.js\2)/)
-    expect(itemShared).not.toMatch(/subpackages_item.*subpackages_user\/common\.js/)
-    expect(userShared).not.toMatch(/subpackages_item.*subpackages_user\/common\.js/)
-    expect(itemVendorRuntimePath).toBeDefined()
-    expect(userVendorRuntimePath).toBeDefined()
-    expect(itemVendorRuntimePath).toContain(`${path.sep}weapp-vendors${path.sep}`)
-    expect(userVendorRuntimePath).toContain(`${path.sep}weapp-vendors${path.sep}`)
-
-    expect(itemSubPackage?.pages).toEqual([
-      'index',
-      'issue-769/index',
-      'login-required/index',
-    ])
-    expect(await fs.pathExists(invalidSharedPageWxmlPath)).toBe(false)
-    expect(await fs.pathExists(itemInvalidCommonPath)).toBe(false)
-    expect(await fs.pathExists(userInvalidCommonPath)).toBe(false)
-    expect(await fs.pathExists(itemVendorsPath)).toBe(false)
-    expect(await fs.pathExists(userVendorsPath)).toBe(false)
-    expect(await fs.pathExists(itemRuntimePath)).toBe(true)
-    expect(await fs.pathExists(userRuntimePath)).toBe(true)
-  })
-
-  it('issue #327: routes npm deps by mainPackage/subPackages config and only emits configured main-package npm output', async () => {
-    await runBuild()
-
-    const issue327NpmRoot = path.join(DIST_ROOT, 'subpackages/issue-327/miniprogram_npm')
-    const itemNpmRoot = path.join(DIST_ROOT, 'subpackages/item/miniprogram_npm')
-    const userNpmRoot = path.join(DIST_ROOT, 'subpackages/user/miniprogram_npm')
-    const subpackageDayjsPath = path.join(issue327NpmRoot, 'dayjs/index.js')
-    const subpackageTdesignPath = path.join(issue327NpmRoot, 'tdesign-miniprogram/button/button.js')
-    const itemCamelCasePath = path.join(itemNpmRoot, 'camelcase/index.js')
-    const userMergePath = path.join(userNpmRoot, 'merge/index.js')
-    const mainDayjsPath = path.join(DIST_ROOT, 'miniprogram_npm/dayjs/index.js')
-    const mainTdesignPath = path.join(DIST_ROOT, 'miniprogram_npm/tdesign-miniprogram/button/button.js')
-    const mainLodashPath = path.join(DIST_ROOT, 'miniprogram_npm/camelcase/index.js')
-    const mainMergePath = path.join(DIST_ROOT, 'miniprogram_npm/merge/index.js')
-    const issuePageJsPath = path.join(DIST_ROOT, 'subpackages/issue-327/index.js')
-    const issuePageJsonPath = path.join(DIST_ROOT, 'subpackages/issue-327/index.json')
-    const itemPageJsPath = path.join(DIST_ROOT, 'subpackages/item/index.js')
-    const userPageJsPath = path.join(DIST_ROOT, 'subpackages/user/index.js')
-
-    expect(await fs.pathExists(subpackageDayjsPath)).toBe(true)
-    expect(await fs.pathExists(subpackageTdesignPath)).toBe(true)
-    expect(await fs.pathExists(itemCamelCasePath)).toBe(true)
-    expect(await fs.pathExists(userMergePath)).toBe(true)
-    expect(await fs.pathExists(mainDayjsPath)).toBe(false)
-    expect(await fs.pathExists(mainTdesignPath)).toBe(true)
-    expect(await fs.pathExists(mainLodashPath)).toBe(false)
-    expect(await fs.pathExists(mainMergePath)).toBe(false)
-    expect(await fs.pathExists(path.join(DIST_ROOT, 'miniprogram_npm'))).toBe(true)
-
-    expect(await fs.pathExists(path.join(issue327NpmRoot, 'camelcase/index.js'))).toBe(false)
-    expect(await fs.pathExists(path.join(issue327NpmRoot, 'merge/index.js'))).toBe(false)
-    expect(await fs.pathExists(path.join(itemNpmRoot, 'merge/index.js'))).toBe(false)
-    expect(await fs.pathExists(path.join(itemNpmRoot, 'dayjs/index.js'))).toBe(false)
-    expect(await fs.pathExists(path.join(userNpmRoot, 'camelcase/index.js'))).toBe(false)
-    expect(await fs.pathExists(path.join(userNpmRoot, 'dayjs/index.js'))).toBe(false)
-
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-    const issuePageJson = await fs.readFile(issuePageJsonPath, 'utf-8')
-    const itemPageJs = await fs.readFile(itemPageJsPath, 'utf-8')
-    const userPageJs = await fs.readFile(userPageJsPath, 'utf-8')
-
-    expect(issuePageJs).toContain('dayjs')
-    expect(issuePageJson).toContain('"t-button": "tdesign-miniprogram/button/button"')
-    expect(itemPageJs).toContain('camelcase')
-    expect(itemPageJs).toContain('npmMarker')
-    expect(userPageJs).toContain('merge')
-    expect(userPageJs).toContain('npmMarker')
-  })
-
-  it('issue #300: keeps boolean props available in runtime call-expression bindings', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-300/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-300/index.js')
-    const probeWxmlPath = path.join(DIST_ROOT, 'components/issue-300/PropsDestructureProbe/index.wxml')
-    const probeJsPath = path.join(DIST_ROOT, 'components/issue-300/PropsDestructureProbe/index.js')
-    const strictProbeWxmlPath = path.join(DIST_ROOT, 'components/issue-300/StrictNoPropsVarProbe/index.wxml')
-    const strictProbeJsPath = path.join(DIST_ROOT, 'components/issue-300/StrictNoPropsVarProbe/index.js')
-    const strictProbeJsonPath = path.join(DIST_ROOT, 'components/issue-300/StrictNoPropsVarProbe/index.json')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-    const probeWxml = await fs.readFile(probeWxmlPath, 'utf-8')
-    const probeJs = await fs.readFile(probeJsPath, 'utf-8')
-    const strictProbeWxml = await fs.readFile(strictProbeWxmlPath, 'utf-8')
-    const strictProbeJs = await fs.readFile(strictProbeJsPath, 'utf-8')
-    const strictProbeJson = await fs.readFile(strictProbeJsonPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-300 props destructure boolean binding')
-    expect(issuePageWxml).toContain('toggle bool:')
-    expect(issuePageWxml).toContain('toggle str:')
-    expect(issuePageWxml).toContain('sync toggle props in place')
-    expect(issuePageWxml).toContain('issue300-toggle-bool')
-    expect(issuePageWxml).toContain('issue300-toggle-str')
-    expect(issuePageWxml).toContain('issue300-toggle-sync')
-    expect(issuePageWxml).toContain('strict-no-props-var')
-    expect(issuePageWxml).toContain('primitive-ref-source')
-    expect(issuePageWxml).toContain('ref-object-source')
-    expect(issuePageWxml).toContain('reactive-object-source')
-    expect(issuePageWxml).toContain('case-id="primitive"')
-    expect(issuePageWxml).toContain('case-id="ref-object"')
-    expect(issuePageWxml).toContain('case-id="reactive-object"')
-    expect(issuePageJs).toContain('toggleBool')
-    expect(issuePageJs).toContain('toggleStr')
-    expect(issuePageJs).toContain('syncTogglePropsInPlace')
-    expect(issuePageJs).toContain('_resetE2E')
-    expect(issuePageJs).toContain('_runE2E')
-
-    const probeBindTokens = probeWxml.match(/__wv_bind_\d+/g) ?? []
-    expect(new Set(probeBindTokens).size).toBeGreaterThanOrEqual(2)
-    expect(probeWxml).toMatch(/data-destructured-bool="\{\{__wv_bind_\d+\}\}"/)
-    expect(probeWxml).toMatch(/data-props-bool="\{\{__wv_bind_\d+\}\}"/)
-    expect(probeWxml).not.toContain('String(bool)')
-    expect(probeWxml).not.toContain('String(props.bool)')
-    expect(probeJs).toContain('boolTextFromProps')
-    expect(probeJs).toContain('boolTextFromDestructured')
-    expect(probeJs).toContain('allowNullPropInput: true')
-    expect(probeJs).toContain('__wevuPropsDerivedKeys: [')
-    expect(probeJs).not.toContain('__wevuProps.props')
-
-    expect(strictProbeWxml).toMatch(/\{\{__wv_bind_\d+\}\}/)
-    expect(strictProbeWxml).toMatch(/data-strict-bool="\{\{__wv_bind_\d+\}\}"/)
-    expect(strictProbeWxml).toContain('data-strict-str="{{str}}"')
-    expect(strictProbeWxml).not.toContain('String(bool)')
-    expect(strictProbeJs).toContain('allowNullPropInput: true')
-    expect(strictProbeJs).toContain('__wevuPropsDerivedKeys: [')
-    expect(strictProbeJs).not.toContain('__wevuProps.props')
-    expect(JSON.parse(strictProbeJson)?.styleIsolation).toBe('apply-shared')
-  })
-
-  it('issue #328: keeps setup ref props available on first paint for child string props', async () => {
-    await runBuild()
-
-    const issuePageWxmlPath = path.join(DIST_ROOT, 'pages/issue-328/index.wxml')
-    const issuePageJsPath = path.join(DIST_ROOT, 'pages/issue-328/index.js')
-    const probeWxmlPath = path.join(DIST_ROOT, 'components/issue-328/ValueProbe/index.wxml')
-    const probeJsPath = path.join(DIST_ROOT, 'components/issue-328/ValueProbe/index.js')
-
-    const issuePageWxml = await fs.readFile(issuePageWxmlPath, 'utf-8')
-    const issuePageJs = await fs.readFile(issuePageJsPath, 'utf-8')
-    const probeWxml = await fs.readFile(probeWxmlPath, 'utf-8')
-    const probeJs = await fs.readFile(probeJsPath, 'utf-8')
-
-    expect(issuePageWxml).toContain('issue-328 setup ref prop first paint')
-    expect(issuePageWxml).toContain('value-probe')
-    expect(issuePageWxml).toContain('issue328-toggle')
-    expect(issuePageWxml).toContain('toggle value: {{value1}}')
-    expect(issuePageWxml).toContain('value="{{value1}}"')
-    expect(issuePageJs).toContain('value1')
-    expect(issuePageJs).toContain('advanceValue')
-    expect(issuePageJs).toContain('_runE2E')
-    expect(issuePageJs).toContain('111')
-    expect(issuePageJs).toContain('222')
-
-    expect(probeWxml).toContain('data-current-value="{{props.value}}"')
-    expect(probeWxml).toContain('data-history="{{historyText}}"')
-    expect(probeWxml).toContain('{{historyText}}')
-    expect(probeJs).toContain('valueHistory')
-    expect(probeJs).toContain('historyText')
-    expect(probeJs).toMatch(/default:\s*['"`]0\.00['"`]/)
-  })
-
-  it('issue #373: compiles shared store computed bindings across reLaunch pages', async () => {
-    await runBuild()
-
-    const launchPageWxmlPath = path.join(DIST_ROOT, 'pages/issue-373/launch/index.wxml')
-    const launchPageJsPath = path.join(DIST_ROOT, 'pages/issue-373/launch/index.js')
-    const resultPageWxmlPath = path.join(DIST_ROOT, 'pages/issue-373/result/index.wxml')
-    const resultPageJsPath = path.join(DIST_ROOT, 'pages/issue-373/result/index.js')
-
-    const launchPageWxml = await fs.readFile(launchPageWxmlPath, 'utf-8')
-    const launchPageJs = await fs.readFile(launchPageJsPath, 'utf-8')
-    const resultPageWxml = await fs.readFile(resultPageWxmlPath, 'utf-8')
-    const resultPageJs = await fs.readFile(resultPageJsPath, 'utf-8')
-    const { code: wevuRuntime } = await findWevuVendorChunkContaining(DIST_ROOT, ['useIssue373Store'])
-
-    expect(launchPageWxml).toContain('issue-373 store computed survives reLaunch')
-    expect(launchPageWxml).toContain('launch count: {{count}}')
-    expect(launchPageWxml).toContain('launch doubled: {{doubled}}')
-    expect(launchPageWxml).toContain('data-doubled="{{doubled}}"')
-    expect(launchPageJs).toContain('runRelaunch')
-    expect(launchPageJs).toContain('_runE2E')
-
-    expect(resultPageWxml).toContain('issue-373 reLaunch store computed result')
-    expect(resultPageWxml).toContain('result count: {{count}}')
-    expect(resultPageWxml).toContain('result doubled: {{doubled}}')
-    expect(resultPageWxml).toContain('data-doubled="{{doubled}}"')
-    expect(resultPageJs).toContain('increment')
-    expect(resultPageJs).toContain('_runE2E')
-
-    expect(wevuRuntime).toContain('issue-373-store')
-    expect(launchPageJs).toMatch(/const\s+\{\s*count,\s*doubled\s*\}\s*=\s*[\w$]+\.[\w$]+\(\s*store\s*\)/)
-    expect(resultPageJs).toMatch(/const\s+\{\s*count,\s*doubled\s*\}\s*=\s*[\w$]+\.[\w$]+\(\s*store\s*\)/)
-  })
-
-  it('emits callback, Promise, and native import targets inside their subpackage', async () => {
-    await runBuild()
-
-    const appJson = await fs.readJSON(path.join(DIST_ROOT, 'app.json')) as {
-      subPackages?: Array<{ pages?: string[], root?: string }>
-    }
-    const pageJs = await fs.readFile(path.join(DIST_ROOT, 'pages/require-async/index.js'), 'utf-8')
-    const callbackJs = await fs.readFile(path.join(DIST_ROOT, 'subpackages/require-async/callback.js'), 'utf-8')
-    const promiseJs = await fs.readFile(path.join(DIST_ROOT, 'subpackages/require-async/promise.js'), 'utf-8')
-    const nativeJs = await fs.readFile(path.join(DIST_ROOT, 'subpackages/require-async/import-native.js'), 'utf-8')
-    const outputFiles = await scanFiles(DIST_ROOT)
-    const jsFiles = outputFiles.filter(file => file.endsWith('.js'))
-    const outputCode = new Map(await Promise.all(jsFiles.map(async file => [
-      file,
-      await fs.readFile(path.join(DIST_ROOT, file), 'utf-8'),
-    ] as const)))
-
-    expect(appJson.subPackages).toContainEqual(expect.objectContaining({
-      pages: ['index'],
-      root: 'subpackages/require-async',
-    }))
-    expect(pageJs).toMatch(/require\.async\((['"])\.\.\/\.\.\/subpackages\/require-async\/callback\.js\1\)\.then\(/)
-    expect(pageJs).toMatch(/require\.async\((['"])\.\.\/\.\.\/subpackages\/require-async\/promise\.js\1\)/)
-    expect(pageJs).toMatch(/require\.async\((['"])\.\.\/\.\.\/subpackages\/require-async\/import-native\.js\1\)/)
-    expect(pageJs).not.toMatch(/import\((['"])\.\.\/\.\.\/subpackages\/require-async\/import-native/)
-    expect(callbackJs).toContain('require-async:callback')
-    expect(promiseJs).toContain('require-async:promise')
-    expect(nativeJs).toContain('require-async:native-default')
-    expect(nativeJs).toContain('require-async:native-named')
-
-    for (const marker of [
-      'require-async:native-default',
-      'require-async:native-named',
-      'require-async:transitive',
-    ]) {
-      const markerFiles = [...outputCode]
-        .filter(([, code]) => code.includes(marker))
-        .map(([file]) => file)
-      expect(markerFiles.length).toBeGreaterThan(0)
-      expect(markerFiles.every(file => file.startsWith('subpackages/require-async/'))).toBe(true)
-    }
-
-    expect(await fs.pathExists(path.join(DIST_ROOT, 'callback.js'))).toBe(false)
-    expect(await fs.pathExists(path.join(DIST_ROOT, 'import-native.js'))).toBe(false)
-    expect(await fs.pathExists(path.join(DIST_ROOT, 'transitive.js'))).toBe(false)
-  })
+  registerGithubIssuesBuildCases(createGithubIssuesBuildCaseContext())
+  registerGithubIssuesBuildLegacyCases()
 })
