+++ e2e/ci/githubIssuesBuild/cases/issue838.config.ts
@@ -1,0 +1,16 @@
+import { mergeConfig } from 'vite'
+import { defineConfig } from 'weapp-vite'
+import baseConfig from '../../../../e2e-apps/github-issues/weapp-vite.config'
+
+export default defineConfig(mergeConfig(baseConfig, {
+  build: {
+    minify: false,
+    outDir: 'dist-issue-838',
+  },
+  weapp: {
+    styles: {
+      source: 'styles/issue-838-app-vue.scss',
+      include: 'app.vue',
+    },
+  },
+}))
