+++ e2e/ci/githubIssuesBuild/cases/types.ts
@@ -1,0 +1,10 @@
+export interface GithubIssuesBuildCaseContext {
+  appRoot: string
+  cliPath: string
+  distRoot: string
+  runStandardBuild: () => Promise<void>
+}
+
+export interface GithubIssuesBuildCaseModule {
+  registerGithubIssuesBuildCase: (context: GithubIssuesBuildCaseContext) => void
+}
